@@ -3468,7 +3468,7 @@ $messages['pt-br'] = [
 	'article-comments-show-all' => 'Mostrar todos os comentários',
 	'article-comments-prev-page' => 'Anterior',
 	'article-comments-next-page' => 'Próximo',
-	'article-comments-page-spacer' => '& # 160...&#160;',
+	'article-comments-page-spacer' => '&#160;...&#160;',
 	'article-comments-delete-reason' => 'O artigo raiz / comentário raiz foi deletado.',
 	'article-comments-empty-comment' => 'Você não pode postar um comentário vazio. [$1 Excluí-lo em vez disso?]',
 	'article-comments-show-more' => 'Mostrar mais comentários',
