@@ -1,18 +1,5 @@
 <?php
 
-/**
- * ArticleComments
- *
- * A ArticleComments extension for MediaWiki
- * Adding comment functionality on article pages
- *
- * @author Jakub Kurcek  <jakub@wikia.inc>
- * @copyright Copyright (C) 2010 Jakub Kurcek, Wikia Inc.
- * @license http://www.gnu.org/copyleft/gpl.html GNU General Public License 2.0 or later
- * @package MediaWiki
- *
- */
-
 class CategoryExhibitionAjax {
 
 	static public function axGetArticlesPage() {
@@ -32,27 +19,31 @@ class CategoryExhibitionAjax {
 	}
 
 	/**
-	 * Returns section html.
+	 * Get one page of one section of a category listing as the following structure
+	 * [  'page' => HTML of the requested page
+	 *    'paginator' => HTML of the paginator ]
+	 * 'articleId', 'page', 'sort', 'display' params will be read from URL.
 	 *
 	 * @param $class string
-	 * @return string
+	 * @return array
 	 */
-
-	static private function getSection( $class ){
+	static private function getSection( $class ) {
 		global $wgRequest;
-		
-		$pageText = $wgRequest->getVal( 'articleId' );
-		$iPaginatorPosition = (int)$wgRequest->getVal( 'page' );
-		$oCategoryTitle = Title::newFromText( $pageText, NS_CATEGORY );
-		if ( !is_object( $oCategoryTitle ) ){
+
+		$pageTitle = $wgRequest->getVal( 'articleId' );
+		$title = Title::newFromText( $pageTitle, NS_CATEGORY );
+		if ( !is_object( $title ) ) {
 			return '';
 		}
-		$oSection = new $class( $oCategoryTitle );
-		$sUrl = $oCategoryTitle->getFullURL();
 
-		$result = $oSection->getSectionAxHTML( $iPaginatorPosition, $sUrl );
-		
-		return $result;
+		$urlParams = new CategoryUrlParams( $wgRequest );
+
+		/** @var CategoryExhibitionSection $oSection */
+		$oSection = new $class( $title, $urlParams );
+		$results = $oSection->getTemplateVars();
+		return [
+			'page' => $results['items'],
+			'paginator' => $results['paginator'],
+		];
 	}
-
 }
