@@ -4,24 +4,10 @@
  * Main Category Gallery class
  */
 class CategoryExhibitionSectionBlogs extends CategoryExhibitionSection {
-
-	public $urlParameter = 'blogs'; // contains section url variable that stores pagination
-	public $templateName = 'blogs';
-
-	public function getSectionHTML(){
-
+	protected function generateSectionData() {
 		global $wgCategoryExhibitionBlogsSectionRows;
-		$oTmpl = $this->getTemplateForNameSpace( 500, $wgCategoryExhibitionBlogsSectionRows * 4 );
-		return $this->executeTemplate( $oTmpl );
-	}
-
-	public function getSectionAxHTML( $paginatorPosition, $sUrl ){
-
-		global $wgCategoryExhibitionBlogsSectionRows;
-		$this->isFromAjax = true;
-		$this->paginatorPosition = $paginatorPosition;
-		$this->sUrl = $sUrl;
-		$oTmpl = $this->getTemplateForNameSpace( 500, $wgCategoryExhibitionBlogsSectionRows * 4 );
-		return $this->executeTemplate( $oTmpl );
+		$this->sectionId = 'mw-blogs';
+		$this->headerMessage = wfMessage( 'category-exhibition-blogs-header' );
+		$this->generateData( 500, $wgCategoryExhibitionBlogsSectionRows * 4 );
 	}
 }
