@@ -18,37 +18,30 @@ $wgExtensionCredits['other'][] = array(
 	'url' => 'https://github.com/Wikia/app/tree/dev/extensions/wikia/CategoryExhibition',
 );
 
-// Calculate the base directory and use it later a few times
-$dir = dirname(__FILE__) . '/';
-
-// Autoload Category Page classes
-$wgAutoloadClasses[ 'CategoryExhibitionPage' ]		= $dir . 'CategoryExhibitionPage.php';
-$wgAutoloadClasses[ 'CategoryPageII' ]			= $dir . 'CategoryPageII.php';
-
-// Autoload extension classes
-$wgAutoloadClasses[ 'CategoryExhibitionSection' ]	= $dir . 'CategoryExhibitionSection.class.php';
-$wgAutoloadClasses[ 'CategoryExhibitionHelper' ]	= $dir . 'CategoryExhibitionHelper.class.php';
-$wgAutoloadClasses[ 'CategoryExhibitionAjax' ]		= $dir . 'CategoryExhibitionAjax.class.php';
-
-// Autoload services
-$wgAutoloadClasses[ 'CategoryDataService' ]		= $dir . 'services/CategoryDataService.class.php';
-
-// Autoloads section data
-$wgAutoloadClasses[ 'CategoryExhibitionSectionBlogs' ]	= $dir . 'CategoryExhibitionSectionBlogs.class.php';
-$wgAutoloadClasses[ 'CategoryExhibitionSectionMedia' ]	= $dir . 'CategoryExhibitionSectionMedia.class.php';
-$wgAutoloadClasses[ 'CategoryExhibitionSectionPages' ]	= $dir . 'CategoryExhibitionSectionPages.class.php';
-$wgAutoloadClasses[ 'CategoryExhibitionSectionSubcategories' ]	= $dir . 'CategoryExhibitionSectionSubcategories.class.php';
+// Autoload
+$wgAutoloadClasses[ 'CategoryDataService' ] = __DIR__ . '/services/CategoryDataService.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionAjax' ] = __DIR__ . '/CategoryExhibitionAjax.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionCacheHelper' ] = __DIR__ . '/CategoryExhibitionCacheHelper.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionHooks' ] = __DIR__ . '/CategoryExhibitionHooks.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionPage' ] = __DIR__ . '/CategoryExhibitionPage.php';
+$wgAutoloadClasses[ 'CategoryExhibitionSection' ] = __DIR__ . '/CategoryExhibitionSection.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionSectionBlogs' ] = __DIR__ . '/CategoryExhibitionSectionBlogs.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionSectionMedia' ] = __DIR__ . '/CategoryExhibitionSectionMedia.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionSectionPages' ] = __DIR__ . '/CategoryExhibitionSectionPages.class.php';
+$wgAutoloadClasses[ 'CategoryExhibitionSectionSubcategories' ] = __DIR__ . '/CategoryExhibitionSectionSubcategories.class.php';
+$wgAutoloadClasses[ 'CategoryPageII' ] = __DIR__ . '/CategoryPageII.php';
+$wgAutoloadClasses[ 'CategoryUrlParams' ] = __DIR__ . '/CategoryUrlParams.class.php';
 
 // Internalization
-$wgExtensionMessagesFiles['CategoryPageII'] = $dir . 'i18n/CategoryExhibition.i18n.php';
+$wgExtensionMessagesFiles[ 'CategoryPageII' ] = __DIR__ . '/i18n/CategoryExhibition.i18n.php';
 
 // Hooks
 define('CATEXHIBITION_DISABLED', 'CATEXHIBITION_DISABLED');
-$wgHooks['LanguageGetMagic'][] = 'CategoryExhibitionHelper::onLanguageGetMagic';
-$wgHooks['InternalParseBeforeLinks'][] = 'CategoryExhibitionHelper::onInternalParseBeforeLinks';
-$wgHooks['ArticleFromTitle'][] = 'CategoryExhibitionHelper::onArticleFromTitle';
-$wgHooks['onArticleSaveComplete'][] = 'CategoryExhibitionHelper::onArticleSaveComplete';
-$wgHooks['AfterCategoriesUpdate'][] = 'CategoryExhibitionHelper::onAfterCategoriesUpdate';
+$wgHooks['LanguageGetMagic'][] = 'CategoryExhibitionHooks::onLanguageGetMagic';
+$wgHooks['InternalParseBeforeLinks'][] = 'CategoryExhibitionHooks::onInternalParseBeforeLinks';
+$wgHooks['ArticleFromTitle'][] = 'CategoryExhibitionHooks::onArticleFromTitle';
+$wgHooks['onArticleSaveComplete'][] = 'CategoryExhibitionHooks::onArticleSaveComplete';
+$wgHooks['AfterCategoriesUpdate'][] = 'CategoryExhibitionHooks::onAfterCategoriesUpdate';
 
 // Ajax dispatcher
 $wgAjaxExportList[] = 'CategoryExhibitionAjax';
