@@ -10,32 +10,27 @@ class CloseMyAccountHelper {
 	// Number of days to wait before properly closing account
 	const CLOSE_MY_ACCOUNT_WAIT_PERIOD = 30;
 
+	const REQUEST_CLOSURE_PREF = 'requested-closure-date';
+	const DISABLED_BY_USER_PREF = 'disabled-by-user-request';
+
 	/**
 	 * Set an account to be closed, and log them out
 	 *
-	 * @param  User    $user The user we're scheduling to be closed
+	 * @param User $user The user we're scheduling to be closed
 	 * @return boolean
 	 */
 	public function scheduleCloseAccount( User $user ) {
-		wfProfileIn( __METHOD__ );
 		// Already set?
 		if ( $this->isScheduledForClosure( $user ) ) {
-			wfProfileOut( __METHOD__ );
 			return true;
 		}
 
-		$user->setGlobalFlag( 'requested-closure', true );
-
-		// Temporarily save requested-closure-date as both an attribute and a preference. This
-		// will be changed to just a preference once we complete the migration. See SOC-2185
-		$user->setGlobalAttribute( 'requested-closure-date', wfTimestamp( TS_DB ) );
-		$user->setGlobalPreference( 'requested-closure-date', wfTimestamp( TS_DB ) );
+		$user->setGlobalPreference( self::REQUEST_CLOSURE_PREF, wfTimestamp( TS_DB ) );
 
 		$user->saveSettings();
 
 		$this->track( $user, 'request-closure' );
 
-		wfProfileOut( __METHOD__ );
 		return true;
 	}
 
@@ -46,16 +41,12 @@ class CloseMyAccountHelper {
 	 * generating a token, and emailing the link with the token, so we
 	 * can confirm that they own this account.
 	 *
-	 * @param  User     $user The user account to reactivate
-	 * @param  WikiaApp $app  An instance of WikiaApp
-	 * @return boolean        True if the reactivation was successfully requested,
-	 *                        False otherwise
+	 * @param User $user The user account to reactivate
+	 * @return boolean True if the reactivation was successfully requested, False otherwise
 	 */
-	public function requestReactivation( User $user, $app ) {
-		wfProfileIn( __METHOD__ );
+	public function requestReactivation( User $user ) {
 		// Not scheduled for closure or not email confirmed?
 		if ( !$this->isScheduledForClosure( $user ) || !$user->isEmailConfirmed() ) {
-			wfProfileOut( __METHOD__ );
 			return false;
 		}
 
@@ -63,7 +54,6 @@ class CloseMyAccountHelper {
 
 		$this->track( $user, 'request-reactivation' );
 
-		wfProfileOut( __METHOD__ );
 		return $response->isGood();
 	}
 
@@ -73,49 +63,37 @@ class CloseMyAccountHelper {
 	 * This involves checking the token, checking the user has authenticated
 	 * successfully, and removing the user options.
 	 *
-	 * @param  User    $user The user account to reactivate
-	 * @return boolean       True if the reactivation was successful,
-	 *                       False otherwise
+	 * @param User $user The user account to reactivate
+	 * @return boolean True if the reactivation was successful, False otherwise
 	 */
 	public function reactivateAccount( User $user ) {
-		wfProfileIn( __METHOD__ );
-
 		// Did they request closure or are already disabled?
 		if ( !$this->isScheduledForClosure( $user ) || $this->isClosed( $user ) ) {
-			wfProfileOut( __METHOD__ );
 			return false;
 		}
 
-		$user->setGlobalFlag( 'requested-closure', null );
-
-		// requested-closure-date is temporarily being stored as both an attribute and a preference.
-		// Make sure to delete from both places. This will be changed to just a preference once the
-		// migration is complete. See SOC-2185
-		$user->setGlobalAttribute( 'requested-closure-date', null );
-		$user->setGlobalPreference( 'requested-closure-date', null );
-
+		$user->setGlobalPreference( self::REQUEST_CLOSURE_PREF, null );
 		$user->saveSettings();
 
 		$this->track( $user, 'account-reactivated' );
 
-		wfProfileOut( __METHOD__ );
 		return true;
 	}
 
 	/**
 	 * Get the number of days until the account is closed
 	 *
-	 * @param  User            $user The user account to check
-	 * @return integer|boolean       The number of days remaining, or false
-	 *                               if no value is set
+	 * @param User $user The user account to check
+	 * @return integer|boolean The number of days remaining, or false if no value is set
 	 */
 	public function getDaysUntilClosure( User $user ) {
 		$daysRemaining = false;
-		$requestDate = $user->getGlobalAttribute( 'requested-closure-date' );
+		$requestDate = $user->getGlobalPreference( self::REQUEST_CLOSURE_PREF );
 
 		if ( $requestDate !== null ) {
 			// Number of days remaining until closure
-			$daysRemaining = (int)( self::CLOSE_MY_ACCOUNT_WAIT_PERIOD - floor( ( time() - strtotime( $requestDate ) ) / 86400 ) );
+			$daysSinceRequest = floor( ( time() - strtotime( $requestDate ) ) / 86400 );
+			$daysRemaining = self::CLOSE_MY_ACCOUNT_WAIT_PERIOD - $daysSinceRequest;
 			if ( $daysRemaining < 0 ) {
 				$daysRemaining = 0;
 			}
@@ -126,33 +104,34 @@ class CloseMyAccountHelper {
 	/**
 	 * Check if the the user account is closed
 	 *
-	 * @param  User    $user The user account to check
-	 * @return boolean       True if the account is disabled,
-	 *                       False otherwise
+	 * @param  User $user The user account to check
+	 * @return boolean True if the account is disabled, False otherwise
 	 */
 	public function isClosed( User $user ) {
-		return (bool)$user->getGlobalFlag( 'disabled', false );
+		return ( bool ) $user->getGlobalFlag( 'disabled', false );
 	}
 
 	/**
 	 * Check if the the user account is scheduled to be closed
 	 *
-	 * @param  User    $user The user account to check
-	 * @return boolean       True if the account is scheduled for closure,
-	 *                       False otherwise
+	 * @param User $user The user account to check
+	 * @return boolean True if the account is scheduled for closure, False otherwise
 	 */
 	public function isScheduledForClosure( User $user ) {
-		return (bool)$user->getGlobalFlag( 'requested-closure', false )
-			&& ( $user->getGlobalAttribute( 'requested-closure-date', false ) !== false );
+		return ( bool ) $user->getGlobalPreference( self::REQUEST_CLOSURE_PREF, false );
 	}
 
 	/**
 	 * Track an event
 	 *
-	 * @param  User   $user   User account the event is affecting
-	 * @param  string $action The type of close account event, can be one of
-	 *                        request-closure, request-reactivation, account-reactivated,
-	 *                        account-closed
+	 * @param User $user User account the event is affecting
+	 * @param string $action The type of close account event, can be one of:
+	 *
+	 *     request-closure
+	 *     request-reactivation
+	 *     account-reactivated
+	 *     account-closed
+	 *
 	 * @return void
 	 */
 	public function track( User $user, $action ) {
@@ -171,5 +150,4 @@ class CloseMyAccountHelper {
 
 		$wgUser = $oldUser;
 	}
-
 }
