@@ -2,6 +2,9 @@
 
 require_once __DIR__ . '/../../../../maintenance/Maintenance.php';
 
+use Wikia\DependencyInjection\Injector;
+use Wikia\Service\User\Preferences\PreferenceService;
+
 /**
  * Maintenance script for closing accounts that have been scheduled
  * to be closed
@@ -15,80 +18,80 @@ class CloseMyAccountMaintenance extends Maintenance {
 	}
 
 	public function execute() {
-		global $wgUser, $wgExternalSharedDB;
+		$this->setCurrentUser();
 
-		// Currently, inverted searches (using a column other than up_user) may not
-		// be supported for preferences in the future. However this feature may be
-		// supported for user "flags" if we need to do reverse lookups like these.
-		// See https://wikia-inc.atlassian.net/browse/SERVICES-469.
-		// --drsnyder
+		$users = $this->getUsers();
+		$numClosed = $this->closeAccounts( $users );
 
-		echo "*******************************************************";
-		echo "Warning, inverted searches on 'user_properties' may be deprecated soon. Contact the services team.";
-		echo "*******************************************************";
+		$this->output( $numClosed . " user accounts closed.\n" );
+	}
+
+	public function getUsers() {
+		/** @var PreferenceService $preferenceService */
+		$preferenceService = Injector::getInjector()->get( PreferenceService::class );
+
+		return $preferenceService->findUsersWithGlobalPreferenceValue( CloseMyAccountHelper::REQUEST_CLOSURE_PREF );
+	}
+
+	public function closeAccounts( $users ) {
+		$accountsClosed = 0;
+
+		if ( empty( $users ) ) {
+			return $accountsClosed;
+		}
 
 		$closeAccountHelper = new CloseMyAccountHelper();
 
-		$dbr = wfGetDB( DB_SLAVE, [], $wgExternalSharedDB );
-
-		// Only get users who self-requested account closure
-		$subQuery = $dbr->selectSQLText(
-			'user_properties',
-			[ 'up_user' ],
-			[
-				'up_property' => 'requested-closure-date',
-			],
-			__METHOD__
-		);
-
-		$res = $dbr->select(
-			'user_properties',
-			[ 'up_user' ],
-			[
-				'up_property' => 'requested-closure',
-				'up_value' => 1,
-				'up_user IN (' . $subQuery . ')',
-			],
-			__METHOD__
-		);
-
-		$usersClosed = [];
-
-		$closeReason = 'User requested account closure more than 30 days ago';
-
-		$wgUser = User::newFromName( 'Wikia' );
-
-		foreach ( $res as $row ) {
-			$userObj = User::newFromId( $row->up_user );
+		foreach ( $users as $userId ) {
+			$userObj = User::newFromId( $userId );
 			$daysRemaining = $closeAccountHelper->getDaysUntilClosure( $userObj );
 
 			if ( $daysRemaining === 0 ) {
-				$this->output( "Closing account {$userObj->getName()}...\n" );
-				$statusMsg1 = '';
-				$statusMsg2 = '';
-				$result = EditAccount::closeAccount( $userObj, $closeReason, $statusMsg1, $statusMsg2, /*$keepEmail = */true );
-
-				// Set an option that signifies this user was closed automatically
-				$userObj->setGlobalFlag( 'disabled-by-user-request', true );
-
-				// Cleanup
-				$userObj->setGlobalFlag( 'requested-closure', null );
-
-				// requested-closure-date is temporarily being stored as both an attribute and a preference.
-				// Make sure to delete from both places. This will be changed to just a preference once the
-				// migration is complete. See SOC-2185
-				$userObj->setGlobalAttribute( 'requested-closure-date', null );
-				$userObj->setGlobalPreference( 'requested-closure-date', null );
-
-				$userObj->saveSettings();
+				$success = $this->closeUserAccount( $userObj );
+				if ( !$success ) {
+					continue;
+				}
 
 				$closeAccountHelper->track( $userObj, 'account-closed' );
 
-				$usersClosed[] = $userObj->getName();
+				$accountsClosed++;
+			} else {
+				$this->output( "User {$userObj->getName()} has $daysRemaining days left\n" );
 			}
 		}
 
-		$this->output( count( $usersClosed ) . " user accounts closed.\n" );
+		return $accountsClosed;
+	}
+
+	public function closeUserAccount( User $userObj ) {
+		$this->output( "Closing account {$userObj->getName()} ... " );
+
+		$closeReason = 'User requested account closure more than 30 days ago';
+		$statusMsg1 = '';
+		$statusMsg2 = '';
+		$keepEmail = true;
+		$success = EditAccount::closeAccount( $userObj, $closeReason, $statusMsg1, $statusMsg2, $keepEmail );
+
+		if ( !$success ) {
+			$this->output( "failed\n" );
+			return false;
+		}
+
+		// Set a preference that indicates this user was closed automatically
+		$userObj->setGlobalPreference( CloseMyAccountHelper::DISABLED_BY_USER_PREF, true );
+
+		// Cleanup
+		$userObj->setGlobalPreference( CloseMyAccountHelper::REQUEST_CLOSURE_PREF, null );
+		$userObj->saveSettings();
+
+		$this->output( "success\n" );
+		return true;
+	}
+
+	public function setCurrentUser() {
+		global $wgUser;
+
+		$wgUser = User::newFromName( 'Wikia' );
 	}
 }
 
