@@ -77,8 +77,14 @@ $messages['en'] = array(
 	'communitypage-view-more' => 'View more',
 	'communitypage-help-module-title' => 'Help',
 	'communitypage-help-module-link-page-name' => 'Contents',
+	'communitypage-help-module-edit-page-name' => 'Editing',
+	'communitypage-help-module-add-link-name' => 'Links',
+	'communitypage-help-module-new-page-name' => 'New_page',
 	'communitypage-help-module-link-text' => 'Need help getting started?',
 	'communitypage-policy-module-title' => 'Community Policy',
+	'communitypage-help-edit-page' => 'How to edit a page',
+	'communitypage-help-add-link' => 'How to add a link',
+	'communitypage-help-add-new-page' => 'How to add a new page',
 	'communitypage-policy-module-text' => 'This community has a set of policies that should be followed by all contributors.',
 	'communitypage-policy-module-edit-link-text' => 'Edit',
 	'communitypage-policy-module-link-text' => 'View policy',
@@ -86,6 +92,14 @@ $messages['en'] = array(
 	'communitypage-anon-contrib-header' => '$1 or $2 and make an edit to this wiki to see how you rank in this community!',
 	'communitypage-anon-login' => 'Log in',
 	'communitypage-anon-register' => 'register',
+	'communitypage-first-edit-heading' => 'Congratulations on making your first contribution to this community!',
+	'communitypage-first-edit-subheading' => 'Connect with others and find more ways to help by visiting the Community Page',
+	'communitypage-first-edit-get-started' => 'Continue',
+	'communitypage-first-edit-maybe-later' => 'Maybe Later',
+	'communitypage-todo-module-heading' => 'Community to-do list',
+	'communitypage-todo-module-edit-list' => 'Edit list',
+	'communitypage-todo-module-description' => 'This is an area for admins to suggest pages that need work to help new contributors get started. (Admins - remember to provide details on what needs to be done.)',
+	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
 );
 
 $messages['qqq'] = array(
@@ -165,6 +179,12 @@ $messages['qqq'] = array(
 	'communitypage-help-module-title' => 'Title of Help module',
 	'communitypage-help-module-link-page-name' => 'Name of article from help namespace with help contents (without Help namespace prefix), this message is used in content language',
 	'communitypage-help-module-link-text' => 'Text in link to help contents page asking user does he need help to getting started',
+	'communitypage-help-module-edit-page-name' => 'Name of article from help namespace with help related to editing a page (without Help namespace prefix)',
+	'communitypage-help-module-add-link-name' => 'Name of article from help namespace with help related to adding a new link (without Help namespace prefix)',
+	'communitypage-help-module-new-page-name' => 'Name of article from help namespace with help related to adding new page (without Help namespace prefix)',
+	'communitypage-help-edit-page' => 'Text in link to help asking user how to edit page',
+	'communitypage-help-add-link' => 'Text in link to help asking user how to add a link',
+	'communitypage-help-add-new-page' => 'Text in link to help asking user how to add new page',
 	'communitypage-policy-module-title' => 'Heading for Community Policy module',
 	'communitypage-policy-module-text' => 'Text describing what the community policy is',
 	'communitypage-policy-module-edit-link-text' => 'Link to edit the community policy',
@@ -173,6 +193,14 @@ $messages['qqq'] = array(
 	'communitypage-anon-contrib-header' => 'Text to show for anonymous users in the heading for top contributors. Tells users to login or register',
 	'communitypage-anon-login' => 'Log in text to use in the context of communitypage-anon-contrib-header $1 together with url',
 	'communitypage-anon-register' => 'register text to use in the context of communitypage-anon-contrib-header $2 together with url',
+	'communitypage-first-edit-heading' => 'Heading for modal dialog shown when user makes his first edit on a community (first edit modal)',
+	'communitypage-first-edit-subheading' => 'Subheading for modal dialog shown when user makes his first edit on a community',
+	'communitypage-first-edit-get-started' => 'Button that leads to community page for first edit modal',
+	'communitypage-first-edit-maybe-later' => 'Button that dismisses the first edit modal',
+	'communitypage-todo-module-heading' => 'Heading for Community To-do list',
+	'communitypage-todo-module-edit-list' => 'Edit list link to show to the right of community to-do list header',
+	'communitypage-todo-module-description' => 'Description of what the purpose of the community to-do list is',
+	'communitypage-todo-module-zero-state' => 'Zero text to show when there is no community to-do page',
 );
 
 $messages['de'] = array(
@@ -205,7 +233,7 @@ $messages['de'] = array(
 	'communitypage-contributions' => '$1 {{PLURAL:$1|Beitrag|Beiträge}}',
 	'communitypage-admins' => 'Admins',
 	'communitypage-other-admins' => 'Andere Admins',
-	'communitypage-recently-joined' => 'Neue Beitragende',
+	'communitypage-recently-joined' => 'Neue Autoren',
 	'communitypage-recent-activity' => 'Alle neuen Aktivitäten anzeigen',
 	'communitypage-all-members' => 'Alle Mitglieder',
 	'communitypage-view-all-members' => 'Alle Autoren anzeigen',
@@ -261,11 +289,21 @@ $messages['de'] = array(
 	'communitypage-anon-login' => 'Anmelden',
 	'communitypage-anon-register' => 'Anmelden',
 	'communitypage-todo-module-page-name' => 'ToDo-Liste der Community',
-	'communitypage-todo-module-heading' => 'ToDo-Liste der Community',
+	'communitypage-todo-module-heading' => 'Augabenliste der Community',
 	'communitypage-todo-module-edit-list' => 'Liste bearbeiten',
 	'communitypage-todo-module-description' => 'In diesem Bereich können Admins Seiten vorschlagen, an denen gearbeitet werden sollte und somit neue Beitragende bei ihrem Start unterstützen. (Admins: denkt bitte daran, detailliert zu beschreiben, was bearbeitet werden sollte).',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => 'Möchtest du gerne erfahren, wie du etwas zu dieser Community beitragen kannst? Kontaktiere einen Admin und bitte ihn, hier Anregungen einzutragen.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Kontaktiere einen Admin',
+	'communitypage-help-module-edit-page-name' => 'Bearbeiten',
+	'communitypage-help-module-add-link-name' => 'Links',
+	'communitypage-help-module-new-page-name' => 'Neue Seite',
+	'communitypage-help-edit-page' => 'Wie man eine Seite bearbeitet',
+	'communitypage-help-add-link' => 'Wie man einen Link einfügt',
+	'communitypage-help-add-new-page' => 'Wie man eine neue Seite erstellt',
+	'communitypage-first-edit-heading' => 'Herzlichen Glückwunsch zu deiner ersten Bearbeitung in dieser Community!',
+	'communitypage-first-edit-subheading' => 'Besuche die Community-Seite um andere Benutzer zu treffen und um herauszufinden, wie du der Community helfen kannst',
+	'communitypage-first-edit-get-started' => 'Leg los',
+	'communitypage-first-edit-maybe-later' => 'Vielleicht später',
 );
 
 $messages['es'] = array(
@@ -357,8 +395,18 @@ $messages['es'] = array(
 	'communitypage-todo-module-heading' => 'Lista de tareas de la Comunidad',
 	'communitypage-todo-module-edit-list' => 'Editar lista',
 	'communitypage-todo-module-description' => 'Esta es un área para que los administradores sugieran páginas que necesitan trabajo para ayudar a los nuevos editores a comenzar. (Administradores - recuerden proporcionar detalles sobre lo que necesita hacerse.)',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => '¿Quieres saber cómo puedes contribuir a esta comunidad? Contacta a los administradores y pídeles que aporten sugerencias aquí.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Contacta a un administrador',
+	'communitypage-help-module-edit-page-name' => 'Editando',
+	'communitypage-help-module-add-link-name' => 'Enlaces',
+	'communitypage-help-module-new-page-name' => 'Nueva_página',
+	'communitypage-help-edit-page' => 'Cómo editar una página',
+	'communitypage-help-add-link' => 'Cómo añadir un enlace',
+	'communitypage-help-add-new-page' => 'Cómo añadir una nueva página',
+	'communitypage-first-edit-heading' => '¡Felicitaciones por haber hecho tu primera edición en esta comunidad!',
+	'communitypage-first-edit-subheading' => 'Conoce a otras personas y encuentra más formas de ayudar visitando la Página Comunitaria',
+	'communitypage-first-edit-get-started' => 'Continuar',
+	'communitypage-first-edit-maybe-later' => 'Tal vez más tarde',
 );
 
 $messages['fr'] = array(
@@ -450,38 +498,48 @@ $messages['fr'] = array(
 	'communitypage-todo-module-heading' => 'À faire dans la communauté',
 	'communitypage-todo-module-edit-list' => 'Modifier la liste',
 	'communitypage-todo-module-description' => "Les administrateurs peuvent ici suggérer des pages à améliorer, sur lesquelles les nouveaux contributeurs peuvent commencer. (Administrateurs, n'oubliez pas de fournir des détails sur ce qu'il faut faire.)",
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => 'Vous voulez savoir comment contribuer à cette communauté ? Contactez un admin afin de lui demander de fournir des suggestions ici.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Contactez un administrateur',
+	'communitypage-help-module-edit-page-name' => 'Modification',
+	'communitypage-help-module-add-link-name' => 'Liens',
+	'communitypage-help-module-new-page-name' => 'Nouvelle_page',
+	'communitypage-help-edit-page' => "Modification d'une page",
+	'communitypage-help-add-link' => "Ajout d'un lien",
+	'communitypage-help-add-new-page' => "Ajout d'une nouvelle page",
+	'communitypage-first-edit-heading' => 'Félicitations pour votre première contribution à cette communauté !',
+	'communitypage-first-edit-subheading' => "Rencontrez d'autres contributeurs et découvrez comment vous pouvez aider en allant sur la Page Communauté",
+	'communitypage-first-edit-get-started' => 'Commencer',
+	'communitypage-first-edit-maybe-later' => 'Une prochaine fois',
 );
 
 $messages['it'] = array(
 	'communitypage-title' => 'Community',
 	'communitypage-header-welcome' => 'Benvenuto nella comunità di {{SITENAME}}!',
 	'communitypage-header-join-button' => 'Iscriviti ora',
-	'communitypage-admin-welcome-message' => 'Benvenuto a {{SITENAME}}. Aiutaci a far crescere questa wiki modificando e aggiungendo contenuti. Apprezziamo i tuoi contributi! Torna qui spesso per conoscere gli ultimi aggiornamenti e informazioni su questa community.',
+	'communitypage-admin-welcome-message' => 'Benvenuto su {{SITENAME}}. Aiutaci a far crescere questa wiki modificando e aggiungendo contenuti. Apprezziamo i tuoi contributi! Torna qui spesso per conoscere gli ultimi aggiornamenti e informazioni su questa community.',
 	'communitypage-admin-group-name' => 'Amministratori',
 	'communitypage-top-contributors' => 'Principali collaboratori',
-	'communitypage-top-contributors-more' => '+ $1 altro {{PLURAL:$1|contributor|contributors}}',
+	'communitypage-top-contributors-more' => '+ $1 {{PLURAL:$1|altro contributo|altri contributi}}',
 	'communitypage-page-list-empty' => 'Nessuna pagina trovata!',
 	'communitypage-page-list-edit' => 'Modifica',
 	'communitypage-page-list-create' => 'Crea',
-	'communitypage-entry-join' => 'Iscriviti a questa community!',
+	'communitypage-entry-join' => 'Unisciti a questa community!',
 	'communitypage-entry-learn-more' => 'Ulteriori informazioni',
 	'communitypage-entry-button' => 'Pagina della community',
-	'communitypage-tasks-title' => 'Attività',
+	'communitypage-tasks-title' => 'Compiti',
 	'communitypage-entry-tasks' => 'Aiutaci a modificare questa wiki',
 	'communitypage-tasks-header-welcome' => 'Aiutaci a far crescere {{SITENAME}}!',
-	'communitypage-tasks-admin-welcome' => 'Benvenuto a {{SITENAME}}. Questa wiki necessita dei tuoi contributi e del tuo aiuto. Ecco alcune pagine di esempio che potrebbero beneficiare delle tue modifiche. Buone modifiche!',
+	'communitypage-tasks-admin-welcome' => 'Benvenuto su {{SITENAME}}. Questa wiki necessita dei tuoi contributi e del tuo aiuto. Ecco alcune pagine di esempio che potrebbero beneficiare delle tue modifiche. Buon lavoro!',
 	'communitypage-invite-friends' => 'Invita gli amici',
 	'communitypage-this-month' => 'Questo mese',
 	'communitypage-pages' => 'Pagine',
-	'communitypage-pageviews' => 'Pageviews',
+	'communitypage-pageviews' => 'Visualizzazioni',
 	'communitypage-edits' => 'Modifiche',
-	'communitypage-editors' => 'Redattori',
-	'communitypage-top-contributors-week' => 'Principali collaboratori questa settimana',
+	'communitypage-editors' => 'Utenti',
+	'communitypage-top-contributors-week' => 'Principali collaboratori di questa settimana',
 	'communitypage-user-rank' => 'Il tuo rank',
-	'communitypage-user-contributions' => '{{PLURAL:$1|contribution|contributions}}',
-	'communitypage-contributions' => '$1 {{PLURAL:$1|contribution|contributions}}',
+	'communitypage-user-contributions' => '{{PLURAL:$1|contributo|contributi}}',
+	'communitypage-contributions' => '$1 {{PLURAL:$1|contributo|contributi}}',
 	'communitypage-admins' => 'Amministratori',
 	'communitypage-other-admins' => 'Altri amministratori',
 	'communitypage-recently-joined' => 'Nuovi collaboratori recenti',
@@ -489,17 +547,17 @@ $messages['it'] = array(
 	'communitypage-all-members' => 'Tutti i membri',
 	'communitypage-view-all-members' => 'Visualizza tutti i collaboratori',
 	'communitypage-activity' => '$1 $2 la pagina $3',
-	'communitypage-created' => 'creato',
-	'communitypage-edited' => 'modificato',
-	'communitypage-deleted' => 'eliminato',
+	'communitypage-created' => 'ha creato',
+	'communitypage-edited' => 'ha modificato',
+	'communitypage-deleted' => 'ha cancellato',
 	'communitypage-admin' => 'Amministratori',
-	'communitypage-joined' => 'è entrato',
+	'communitypage-joined' => 'Si è unito',
 	'communitypage-modal-title' => 'Tutti i collaboratori',
-	'communitypage-modal-tab-all' => 'Tutto',
+	'communitypage-modal-tab-all' => 'Tutti',
 	'communitypage-modal-tab-admins' => 'Amministratori',
-	'communitypage-modal-tab-loadingerror' => 'Errore caricamento dati',
+	'communitypage-modal-tab-loadingerror' => 'Errore nel caricamento dei dati',
 	'communitypage-no-contributions' => 'Non ci sono ancora contributi questa settimana! Sii il primo a contribuire!',
-	'communitypage-no-admins' => 'Non ci sono ancora amministratori in questo wiki.',
+	'communitypage-no-admins' => 'Non ci sono ancora amministratori su questa wiki.',
 	'communitypage-no-admins-contact' => 'Contattaci subito per essere il primo!',
 	'communitypage-communitycentral-link' => 'http://community.wikia.com/wiki/Special:Contact/adoption',
 	'communitypage-no-members' => 'Questa community non ha membri',
@@ -510,20 +568,20 @@ $messages['it'] = array(
 	'communitypage-uncategorizedpages-title' => 'Pagine senza categorie',
 	'communitypage-uncategorizedpages-description' => 'Aggiungi le categorie a queste pagine. Una buona categorizzazione è fondamentale per una wiki ben organizzata!',
 	'communitypage-wantedpages-title' => 'Pagine richieste',
-	'communitypage-wantedpages-description' => "Ci sono dei link in rosso che puntano a queste pagine, ma le pagine non esistono! Crea l'articolo o fanne un reindirizzamento all'articolo corretto per aiutare i lettori a navigare nella tua wiki.",
-	'communitypage-lastrevision' => '$1 modificato il $2',
-	'communitypage-noofviews' => '$1 viste',
+	'communitypage-wantedpages-description' => "Ci sono dei link rossi che puntano a queste pagine, ma le pagine non esistono! Crea l'articolo o fanne un reindirizzamento all'articolo corretto per aiutare i lettori a navigare nella tua wiki.",
+	'communitypage-lastrevision' => 'modificato da $1 il $2',
+	'communitypage-noofviews' => '$1 visualizzazioni',
 	'communitypage-full-list' => 'Visualizza elenco completo',
 	'communitypage-modal-tab-all-contribution-header' => 'Ultimo contributo',
 	'communitypage-deadendpages-title' => 'Pagine senza link',
-	'communitypage-deadendpages-description' => 'Queste pagine dovrebbero contenere dei link ad altri articoli sulla tua wiki. Articoli ben collegati tra di loro tramite link aiutano i lettori a scoprire più informazioni sul tuo argomento e a migliorare la loro classificazione nei motori di ricerca.',
+	'communitypage-deadendpages-description' => 'Queste pagine dovrebbero contenere dei link ad altri articoli sulla tua wiki. Articoli ben collegati tra di loro tramite link aiutano i lettori a scoprire più informazioni sul tuo argomento e migliorano la classificazione nei motori di ricerca.',
 	'communitypage-view-more' => 'Mostra di più',
 	'communitypage-policy-module-link-page-name' => 'Linee guida per la community',
-	'communitypage-recent-activity-header' => 'Attività wiki recente',
-	'communitypage-help-module-title' => 'Guida',
-	'communitypage-help-module-link-page-name' => 'Indice',
+	'communitypage-recent-activity-header' => 'Attività recenti della Wiki',
+	'communitypage-help-module-title' => 'Aiuto',
+	'communitypage-help-module-link-page-name' => 'Contenuti',
 	'communitypage-help-module-link-text' => 'Hai bisogno di aiuto per iniziare?',
-	'communitypage-policy-module-title' => 'Politica sulla community',
+	'communitypage-policy-module-title' => 'Politica della community',
 	'communitypage-policy-module-text' => 'Questa community ha un set di politiche che devono essere seguite da tutti i collaboratori.',
 	'communitypage-policy-module-edit-link-text' => 'Modifica',
 	'communitypage-policy-module-link-text' => 'Visualizza politica',
@@ -533,18 +591,28 @@ $messages['it'] = array(
 	'communitypage-entrypoint-modal-connect-title' => 'Connetti',
 	'communitypage-entrypoint-modal-connect-text' => 'Discuti le tue reazioni, idee e argomenti preferiti con altri fan che amano $1 come te.',
 	'communitypage-entrypoint-modal-explore-title' => 'Esplora',
-	'communitypage-entrypoint-modal-explore-text' => 'Fai parte di qualcosa di grande. La Community $1 è solo una delle 360.000 community presenti nella rete Wikia.',
+	'communitypage-entrypoint-modal-explore-text' => 'Fai parte di qualcosa di grande. La Community $1 è solo una delle 360.000 community presenti nella rete di Wikia.',
 	'communitypage-entrypoint-modal-button-text' => 'Dai una mano',
 	'communitypage-entrypoint-modal-video-button-text' => 'Guarda il video',
-	'communitypage-anon-contrib-header' => '$1 o $2 e apporta una modifica a questo wiki per vedere come ti classifichi in questa community!',
+	'communitypage-anon-contrib-header' => '$1 o $2 e apporta una modifica a questa wiki per vedere come ti classifichi in questa community!',
 	'communitypage-anon-login' => 'Accedi',
-	'communitypage-anon-register' => 'registra',
+	'communitypage-anon-register' => 'registrati',
 	'communitypage-todo-module-page-name' => 'Elenco delle cose da file community',
-	'communitypage-todo-module-heading' => 'Elenco delle cose da fare community',
+	'communitypage-todo-module-heading' => 'Lista delle cose da fare nella community',
 	'communitypage-todo-module-edit-list' => 'Modifica elenco',
-	'communitypage-todo-module-description' => "Si tratta di un'area per gli amministratori per suggerire pagine che necessitano di lavoro, così da aiutare i nuovi collaboratori a iniziare. (Ammin - ricordarsi di fornire dettagli su ciò che occorre fare).",
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-description' => "Questa è un'area per gli amministratori per suggerire pagine che necessitano di lavoro ai nuovi collaboratori in modo da aiutarli ad iniziare. (Amministratori: ricordatevi di fornire dettagli su ciò che occorre fare.)",
+	'communitypage-todo-module-zero-state' => 'Vuoi sapere come puoi contribuire a questa community? Contatta un amministratore e chiedigli di aggiungere dei suggerimenti qui.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Contatta un amministratore',
+	'communitypage-help-module-edit-page-name' => 'Modifica',
+	'communitypage-help-module-add-link-name' => 'Link',
+	'communitypage-help-module-new-page-name' => 'Nuova pagina',
+	'communitypage-help-edit-page' => 'Come modificare una pagina',
+	'communitypage-help-add-link' => 'Come aggiungere un link',
+	'communitypage-help-add-new-page' => 'Come creare una nuova pagina',
+	'communitypage-first-edit-heading' => 'Congratulazioni per aver fatto la tua prima modifica su questa community!',
+	'communitypage-first-edit-subheading' => 'Incontra altra gente e scopri nuovi modi di contribuire visitando la pagina della community',
+	'communitypage-first-edit-get-started' => 'Inizia adesso',
+	'communitypage-first-edit-maybe-later' => 'Forse più tardi',
 );
 
 $messages['ja'] = array(
@@ -561,10 +629,10 @@ $messages['ja'] = array(
 	'communitypage-header-welcome' => '{{SITENAME}} コミュニティへようこそ！',
 	'communitypage-entry-button' => 'コミュニティ・ページ',
 	'communitypage-tasks-title' => 'タスク',
-	'communitypage-entry-tasks' => 'あなたの知識と情熱を共有してください',
+	'communitypage-entry-tasks' => 'あなたの知識と情熱を共有してください！',
 	'communitypage-tasks-admin-welcome' => '{{SITENAME}} へようこそ。このコミュニティはあなたの知識と情熱を求めています。もっと書き込みが必要なページの一部をこちらに紹介しています。まずは知っていることを書き込んでみよう！',
 	'communitypage-communitycentral-link' => 'http://ja.community.wikia.com/wiki/Special:Contact/adoption',
-	'communitypage-admin-welcome-message' => '{{SITENAME}} へようこそ。あなたの知っていることをここに書き込んでみませんか？同じことに興味や関心を持つファン同士がこのコミュニティに集まり、知識と情熱を共有することで、そのつながりはもっと広がるはず！どんな情報が更新されたかもこのページでチェックしよう。',
+	'communitypage-admin-welcome-message' => '{{SITENAME}} へようこそ。あなたの知っていることをここに書き込んでみませんか？ 同じことに興味や関心を持つファン同士がこのコミュニティに集まり、知識と情熱を共有することで、そのつながりはもっと広がるはず！ どんな情報が更新されたかもこのページでチェックしよう。',
 	'communitypage-page-list-create' => '新規作成',
 	'communitypage-invite-friends' => '友だちを招待',
 	'communitypage-this-month' => '今月',
@@ -601,15 +669,15 @@ $messages['ja'] = array(
 	'communitypage-popularpages-title' => '人気のページ',
 	'communitypage-popularpages-description' => 'これらはアクセスの多いページです。最新の情報や画像、より深い知識を書き込んでもっとページを魅力的にしよう！',
 	'communitypage-uncategorizedpages-title' => 'カテゴリがないページ',
-	'communitypage-uncategorizedpages-description' => 'これらのページにカテゴリを追加しましょう。コミュニティのコンテンツを整理するには、適切なカテゴリの追加が不可欠です。',
+	'communitypage-uncategorizedpages-description' => 'これらのページにカテゴリを追加しよう。カテゴリを追加するとコミュニティのコンテンツが整理され、より読者にわかりやすくなります。',
 	'communitypage-wantedpages-title' => '作成が望まれているページ',
-	'communitypage-wantedpages-description' => 'これらの項目は、まだページが存在していません。新たに情報を書き込むか、正しいページにリダイレクトするよう設定し、もっとコミュニティを充実させましょう。',
+	'communitypage-wantedpages-description' => 'これらの項目は、まだページが存在していません。新たに情報を書き込むか、正しいページにリダイレクト(転送)するよう設定し、もっとコミュニティを充実させよう。',
 	'communitypage-lastrevision' => '$2に$1さんが編集しました',
 	'communitypage-noofviews' => '$1件のビュー',
 	'communitypage-full-list' => 'すべて見る',
 	'communitypage-modal-tab-all-contribution-header' => '最終更新',
 	'communitypage-deadendpages-title' => 'リンクがないページ',
-	'communitypage-deadendpages-description' => 'これらのページに、コミュニティ内の別のページへのリンクを張りましょう。関連する内容のページにリンクを張ると読者がより知識を深めることができるようになり、さらに検索エンジンでのランキングが上がることにもつながります。',
+	'communitypage-deadendpages-description' => 'これらのページに、コミュニティ内の別のページへのリンクを張ってみよう。関連する内容のページにリンクを張ると読者がより知識を深めることができるようになり、さらに検索エンジンでのランキングが上がることにもつながります。',
 	'communitypage-view-more' => 'もっと見る',
 	'communitypage-policy-module-link-page-name' => 'コミュニティ・ガイドライン',
 	'communitypage-recent-activity-header' => '最近のコミュニティ更新情報',
@@ -624,7 +692,7 @@ $messages['ja'] = array(
 	'communitypage-entrypoint-modal-edit-title' => '編集',
 	'communitypage-entrypoint-modal-edit-text' => '同じ想いをもつ仲間と一緒に、記事の執筆、さまざまな知識・情報の整理、画像の投稿などを通じて、ファンによるファンのための、究極の$1コミュニティを作り上げよう。',
 	'communitypage-entrypoint-modal-connect-title' => 'つながる',
-	'communitypage-entrypoint-modal-connect-text' => 'あなたのファン活動やアイディア、お気に入りのトピックなどを同じく$1を愛する他のファンと共有し語り合おう。',
+	'communitypage-entrypoint-modal-connect-text' => 'あなたのファン活動やアイディア、お気に入りのトピックなどを、同じく$1を愛する他のファンと共有し語り合おう。',
 	'communitypage-entrypoint-modal-explore-title' => '探索',
 	'communitypage-entrypoint-modal-explore-text' => 'Wikiaには$1コミュニティの他に360,000以上ものコミュニティが活動しています。世界的なファン・ネットワークの一員になり、新たな知識や仲間を見つけよう。',
 	'communitypage-entrypoint-modal-button-text' => '参加する',
@@ -635,112 +703,132 @@ $messages['ja'] = array(
 	'communitypage-todo-module-page-name' => 'このコミュニティのTo-Doリスト',
 	'communitypage-todo-module-heading' => 'このコミュニティのTo-Doリスト',
 	'communitypage-todo-module-edit-list' => 'リストを編集する',
-	'communitypage-todo-module-description' => 'ここではアドミン(管理者)が、読者やほかの編集メンバーに、より良くしたいページを紹介することができます。(アドミンの方へ -ほかのメンバーが編集しやすいように、それぞれのページを改善するためのポイントをできるだけ詳しく記載するようにしてください)',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-description' => 'ここではアドミン(管理者)が、読者やほかの編集メンバーに、より良くしたいページを紹介することができます。(アドミンの方へ - ほかのメンバーが編集しやすいように、それぞれのページを改善するためのポイントをできるだけ詳しく記載するようにしてください。)',
+	'communitypage-todo-module-zero-state' => 'このコミュニティを編集したいと思ったら、どのページがヘルプを求めているかアドミンに連絡して聞いてみよう。',
 	'communitypage-todo-module-zero-state-contact-admin' => 'アドミンに連絡',
+	'communitypage-help-module-edit-page-name' => '編集中',
+	'communitypage-help-module-add-link-name' => 'リンク',
+	'communitypage-help-module-new-page-name' => 'New_page',
+	'communitypage-help-edit-page' => 'ページを編集するには',
+	'communitypage-help-add-link' => 'リンクを追加するには',
+	'communitypage-help-add-new-page' => 'ページを追加するには',
+	'communitypage-first-edit-heading' => 'これでコミュニティへの初投稿は無事完了です！',
+	'communitypage-first-edit-subheading' => 'コミュニティページにアクセスしよう！他のファンとの出会いや、コミュニティへの参加方法についての詳しい情報が見つかります。',
+	'communitypage-first-edit-get-started' => '編集スタート！',
+	'communitypage-first-edit-maybe-later' => '後で',
 );
 
 $messages['nl'] = array(
 	'communitypage-title' => 'Community',
-	'communitypage-header-welcome' => 'Welcome to the {{SITENAME}} community!',
-	'communitypage-header-join-button' => 'Join Now',
-	'communitypage-admin-welcome-message' => 'Welcome to {{SITENAME}}. Help us grow this wiki by editing and adding content. We welcome and appreciate your contributions! Check back here often for the latest updates and information about this community.',
-	'communitypage-admin-group-name' => 'Admin',
-	'communitypage-top-contributors' => 'Top contributors',
-	'communitypage-top-contributors-more' => '+ $1 more {{PLURAL:$1|contributor|contributors}}',
-	'communitypage-page-list-empty' => 'No pages found!',
-	'communitypage-page-list-edit' => 'Edit',
-	'communitypage-page-list-create' => 'Create',
-	'communitypage-entry-join' => 'Join this community!',
-	'communitypage-entry-learn-more' => 'Learn more',
-	'communitypage-entry-button' => 'Community Page',
-	'communitypage-tasks-title' => 'Tasks',
-	'communitypage-entry-tasks' => 'Help us edit this wikia',
-	'communitypage-tasks-header-welcome' => 'Help us grow {{SITENAME}}!',
-	'communitypage-tasks-admin-welcome' => 'Welcome to {{SITENAME}}. This wikia needs your contributions and your help. Here are some example pages that could use your edits. Happy editing!',
-	'communitypage-invite-friends' => 'Invite Friends',
-	'communitypage-this-month' => 'This Month',
-	'communitypage-pages' => 'Pages',
-	'communitypage-pageviews' => 'Pageviews',
-	'communitypage-edits' => 'Edits',
-	'communitypage-editors' => 'Editors',
-	'communitypage-top-contributors-week' => 'Top Contributors This Week',
-	'communitypage-user-rank' => 'Your Rank',
-	'communitypage-user-contributions' => '{{PLURAL:$1|contribution|contributions}}',
-	'communitypage-contributions' => '$1 {{PLURAL:$1|contribution|contributions}}',
-	'communitypage-admins' => 'Admins',
-	'communitypage-other-admins' => 'Other Admins',
-	'communitypage-recently-joined' => 'Recent New Contributors',
-	'communitypage-recent-activity' => 'View All Recent Activity',
-	'communitypage-all-members' => 'All Members',
-	'communitypage-view-all-members' => 'View All Contributors',
-	'communitypage-activity' => '$1 $2 the page $3',
-	'communitypage-created' => 'created',
-	'communitypage-edited' => 'edited',
-	'communitypage-deleted' => 'deleted',
-	'communitypage-admin' => 'Admin',
-	'communitypage-joined' => 'Joined',
-	'communitypage-modal-title' => 'All Contributors',
-	'communitypage-modal-tab-all' => 'All',
-	'communitypage-modal-tab-all-contribution-header' => 'Last contribution',
-	'communitypage-modal-tab-admins' => 'Admins',
-	'communitypage-modal-tab-loadingerror' => 'Error loading data',
-	'communitypage-no-contributions' => 'There are no contributions yet this week! Be the first to contribute!',
-	'communitypage-no-admins' => 'There are no Admins on this wiki yet.',
-	'communitypage-no-admins-contact' => 'Contact us to be the first!',
-	'communitypage-communitycentral-link' => 'http://community.wikia.com/wiki/Special:Contact/adoption',
-	'communitypage-no-members' => 'This community has no members',
-	'communitypage-help-us-grow' => "Help us grow '''{{SITENAME}}'''!",
-	'communitypage-new' => 'New',
-	'communitypage-popularpages-title' => 'Top Rising Pages',
-	'communitypage-popularpages-description' => 'These pages are gaining traction with your readers! Make sure they are up to your standards and have all the latest images and information.',
-	'communitypage-uncategorizedpages-title' => 'Pages without categories',
-	'communitypage-uncategorizedpages-description' => 'Add categories to these pages. Good categorization is vital to a successfully organized wiki!',
-	'communitypage-wantedpages-title' => 'Wanted pages',
-	'communitypage-wantedpages-description' => "There are redlinks pointing to these pages, but the pages don't exist! Create the article or make it a redirect to the correct article to help readers navigate your wiki.",
-	'communitypage-deadendpages-title' => 'Pages without links',
-	'communitypage-deadendpages-description' => 'These pages should link to other articles on your wiki. Well-linked articles help readers discover more information about your topic and improve search engine rankings.',
-	'communitypage-lastrevision' => '$1 edited on $2',
-	'communitypage-noofviews' => '$1 views',
-	'communitypage-full-list' => 'View full list',
-	'communitypage-view-more' => 'View more',
-	'communitypage-recent-activity-header' => 'Recent wiki activity',
+	'communitypage-header-welcome' => 'Welkom op de {{SITENAME}} community!',
+	'communitypage-header-join-button' => 'Nu lid worden',
+	'communitypage-admin-welcome-message' => 'Welkom op {{SITENAME}}. Help ons deze wiki aan te groeien door inhoud te bewerken en toe te voegen. Wij verwelkomen en waarderen je bewerkingen! Kom vaker terug voor de meest recente updates en informatie over deze community.',
+	'communitypage-admin-group-name' => 'Beheerder',
+	'communitypage-top-contributors' => 'Top bewerkers',
+	'communitypage-top-contributors-more' => '+ nog $1 {{PLURAL:$1|bewerker|bewerkers}}',
+	'communitypage-page-list-empty' => "Geen pagina's gevonden!",
+	'communitypage-page-list-edit' => 'Bewerken',
+	'communitypage-page-list-create' => 'Aanmaken',
+	'communitypage-entry-join' => 'Word lid van deze community!',
+	'communitypage-entry-learn-more' => 'Meer te weten komen',
+	'communitypage-entry-button' => 'Community-pagina',
+	'communitypage-tasks-title' => 'Taken',
+	'communitypage-entry-tasks' => 'Help ons deze wikia te bewerken',
+	'communitypage-tasks-header-welcome' => "Help ons om '''{{SITENAME}}''' te doen groeien!",
+	'communitypage-tasks-admin-welcome' => "Welkom op {{SITENAME}}. Deze wikia heeft jouw bijdragen en hulp nodig. Hier zijn enkele voorbeeldpagina's die jouw bewerkingen zouden kunnen gebruiken. Veel plezier!",
+	'communitypage-invite-friends' => 'Vrienden uitnodigen',
+	'communitypage-this-month' => 'Deze maand',
+	'communitypage-pages' => "Pagina 's",
+	'communitypage-pageviews' => 'Weergaven',
+	'communitypage-edits' => 'Bewerkingen',
+	'communitypage-editors' => 'Bewerkers',
+	'communitypage-top-contributors-week' => 'Top bewerkers deze week',
+	'communitypage-user-rank' => 'Jouw rang',
+	'communitypage-user-contributions' => '{{PLURAL:$1|bewerking|bewerkingen}}',
+	'communitypage-contributions' => '$1 {{PLURAL:$1|bewerking|bewerkingen}}',
+	'communitypage-admins' => 'Beheerders',
+	'communitypage-other-admins' => 'Andere beheerders',
+	'communitypage-recently-joined' => 'Recente nieuwe bewerkers',
+	'communitypage-recent-activity' => 'Alle recente activiteit bekijken',
+	'communitypage-all-members' => 'Alle leden',
+	'communitypage-view-all-members' => 'Alle bewerkers bekijken',
+	'communitypage-activity' => '$1 $2 $3',
+	'communitypage-created' => 'maakte',
+	'communitypage-edited' => 'bewerkte',
+	'communitypage-deleted' => 'verwijderde',
+	'communitypage-admin' => 'Beheerder',
+	'communitypage-joined' => 'Lid geworden',
+	'communitypage-modal-title' => 'Alle bewerkers',
+	'communitypage-modal-tab-all' => 'Iedereen',
+	'communitypage-modal-tab-all-contribution-header' => 'Laatste bewerking',
+	'communitypage-modal-tab-admins' => 'Beheerders',
+	'communitypage-modal-tab-loadingerror' => 'Fout tijdens het laden van de gegevens',
+	'communitypage-no-contributions' => 'Er zijn nog geen bewerkingen geweest deze week! Wees de eerste om bij te dragen!',
+	'communitypage-no-admins' => 'Er zijn nog geen beheerders op deze wiki.',
+	'communitypage-no-admins-contact' => 'Contacteer ons om er een te worden!',
+	'communitypage-communitycentral-link' => 'http://nl.community.wikia.com/wiki/Special:Contact/adoption',
+	'communitypage-no-members' => 'Deze community heeft geen leden',
+	'communitypage-help-us-grow' => "Help ons om '''{{SITENAME}}''' te doen groeien!",
+	'communitypage-new' => 'Nieuw',
+	'communitypage-popularpages-title' => "Trending pagina's",
+	'communitypage-popularpages-description' => "Deze pagina's worden populairder bij de lezers! Zorg ervoor dat ze aan je standaarden voldoen en alle nieuwste afbeeldingen en info bevatten.",
+	'communitypage-uncategorizedpages-title' => "Pagina's zonder categorieën",
+	'communitypage-uncategorizedpages-description' => "Voeg categorieën toe aan deze pagina's. Een degelijke categorisatie is essentieel voor een goed georganiseerde wiki!",
+	'communitypage-wantedpages-title' => "Gevraagde pagina's",
+	'communitypage-wantedpages-description' => "Er zijn rode links die naar deze pagina's verwijzen, maar de pagina's bestaan niet! Maak het artikel of maak een doorverwijzing naar het artikel, zodat lezers gemakkelijker je wiki kunnen navigeren.",
+	'communitypage-deadendpages-title' => "Pagina's zonder links",
+	'communitypage-deadendpages-description' => "Deze pagina's zouden moeten linken naar andere artikels op je wiki. Goed verbonden artikels helpen lezers om meer info te ontdekken over het onderwerp en om je rang in zoekmachines te verbeteren.",
+	'communitypage-lastrevision' => '$1 heeft bewerkt op $2',
+	'communitypage-noofviews' => '$1 weergaves',
+	'communitypage-full-list' => 'Volledige lijst weergeven',
+	'communitypage-view-more' => 'Meer weergeven',
+	'communitypage-recent-activity-header' => 'Recente wikiactiviteit',
 	'communitypage-help-module-title' => 'Help',
-	'communitypage-help-module-link-page-name' => 'Contents',
-	'communitypage-help-module-link-text' => 'Need help getting started?',
-	'communitypage-policy-module-title' => 'Community Policy',
-	'communitypage-policy-module-text' => 'This community has a set of policies that should be followed by all contributors.',
-	'communitypage-policy-module-edit-link-text' => 'Edit',
-	'communitypage-policy-module-link-text' => 'View policy',
-	'communitypage-policy-module-link-page-name' => 'Community guidelines',
-	'communitypage-entrypoint-modal-title' => 'The $1 Community is built by $2 fans like you.',
-	'communitypage-entrypoint-modal-edit-title' => 'Edit',
-	'communitypage-entrypoint-modal-edit-text' => 'Help fellow members write articles, organize facts, and upload images to build the ultimate fan-run $1 community.',
-	'communitypage-entrypoint-modal-connect-title' => 'Connect',
-	'communitypage-entrypoint-modal-connect-text' => 'Discuss your reactions, ideas, and favorite topics with other fans who love $1 as much as you do.',
-	'communitypage-entrypoint-modal-explore-title' => 'Explore',
-	'communitypage-entrypoint-modal-explore-text' => 'Be part of something huge. The $1 Community is just one of 360,000 communities on the Wikia network.',
-	'communitypage-entrypoint-modal-button-text' => 'Help out',
-	'communitypage-entrypoint-modal-video-button-text' => 'Watch the video',
-	'communitypage-anon-contrib-header' => '$1 or $2 and make an edit to this wiki to see how you rank in this community!',
-	'communitypage-anon-login' => 'Log in',
-	'communitypage-anon-register' => 'register',
+	'communitypage-help-module-link-page-name' => 'Inhoud',
+	'communitypage-help-module-link-text' => 'Hulp nodig om te beginnen?',
+	'communitypage-policy-module-title' => 'Communitybeleid',
+	'communitypage-policy-module-text' => 'Deze community heeft een aantal beleidsregels die door alle bewerkers gevolgd moeten worden.',
+	'communitypage-policy-module-edit-link-text' => 'Bewerken',
+	'communitypage-policy-module-link-text' => 'Regels weergeven',
+	'communitypage-policy-module-link-page-name' => 'Community richtlijnen',
+	'communitypage-entrypoint-modal-title' => 'De $1 community is opgebouwd door $2-fans zoals jij.',
+	'communitypage-entrypoint-modal-edit-title' => 'Bewerken',
+	'communitypage-entrypoint-modal-edit-text' => 'Help andere leden om artikels te schrijven, feiten te organiseren en afbeeldingen te uploaden om de ultieme door fans gedreven $1 community te bouwen.',
+	'communitypage-entrypoint-modal-connect-title' => 'Samenkomen',
+	'communitypage-entrypoint-modal-connect-text' => 'Bespreek je reacties, ideeën en favoriete onderwerpen met andere fans die net zoveel van $1 houden als jij.',
+	'communitypage-entrypoint-modal-explore-title' => 'Verkennen',
+	'communitypage-entrypoint-modal-explore-text' => "Maak deel uit van iets groots. De $1 community is maar één van de 360 000 community's op het Wikia-netwerk.",
+	'communitypage-entrypoint-modal-button-text' => 'Meehelpen',
+	'communitypage-entrypoint-modal-video-button-text' => 'Video bekijken',
+	'communitypage-anon-contrib-header' => '$1 of $2, en maak een bewerking op deze wiki om jouw rang in de community te zien!',
+	'communitypage-anon-login' => 'Aanmelden',
+	'communitypage-anon-register' => 'registreren',
 	'communitypage-todo-module-page-name' => 'Community-to-do-list',
 	'communitypage-todo-module-heading' => 'Community To-do list',
-	'communitypage-todo-module-edit-list' => 'Edit list',
-	'communitypage-todo-module-description' => 'This is an area for admins to suggest pages that need work to help new contributors get started. (Admins - remember to provide details on what needs to be done.)',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-edit-list' => 'Lijst bewerkingen',
+	'communitypage-todo-module-description' => "Dit is een plaats voor beheerders om pagina's waar nog werk aan is, aan te raden aan nieuwe bewerkers om hen te helpen beginnen. (Beheerders: denk eraan om te beschrijven wat er nog aan gedaan moet worden.)",
+	'communitypage-todo-module-zero-state' => 'Wil je weten hoe je kan bijdragen aan deze community? Contacteer een beheerder en vraag om wat voorstellen hier.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Contact an admin',
+	'communitypage-help-module-edit-page-name' => 'Bewerken',
+	'communitypage-help-module-add-link-name' => 'Koppelingen',
+	'communitypage-help-module-new-page-name' => 'Nieuwe_pagina',
+	'communitypage-help-edit-page' => 'Hoe een pagina te bewerken',
+	'communitypage-help-add-link' => 'Hoe een link toe te voegen',
+	'communitypage-help-add-new-page' => 'Hoe een nieuwe pagina toevoegen',
+	'communitypage-first-edit-heading' => 'Congratulations on making your first edit to this community!',
+	'communitypage-first-edit-subheading' => 'Meet others and find more ways to help by visiting the Community Page',
+	'communitypage-first-edit-get-started' => 'Get Started',
+	'communitypage-first-edit-maybe-later' => 'Maybe Later',
 );
 
 $messages['pl'] = array(
 	'communitypage-title' => 'Społeczność',
 	'communitypage-header-welcome' => 'Witamy w społeczności {{SITENAME}}!',
 	'communitypage-header-join-button' => 'Dołącz teraz',
-	'communitypage-admin-welcome-message' => 'Witaj na {{SITENAME}}. Pomóż nam rozwijać tą wikię poprzez dodawanie i edycję treści. Będziemy wdzięczni za twój wkład! Wracaj tutaj często, aby na bieżąco śledzić zmiany i informacje na temat tej społeczności.',
+	'communitypage-admin-welcome-message' => 'Witaj na {{SITENAME}}. Pomóż nam rozwijać tę wikię poprzez dodawanie i edycję treści. Będziemy wdzięczni za twój wkład! Wracaj tu często, aby na bieżąco śledzić zmiany i informacje na temat społeczności.',
 	'communitypage-admin-group-name' => 'Administrator',
 	'communitypage-top-contributors' => 'Najaktywniejsi edytorzy',
-	'communitypage-top-contributors-more' => '+ $1 więcej {{PLURAL:$1|współtwórca|współtwórców}}',
+	'communitypage-top-contributors-more' => '+ $1 {{PLURAL:$1|współtwórca|współtwórców}} więcej',
 	'communitypage-page-list-empty' => 'Nie odnaleziono stron!',
 	'communitypage-page-list-edit' => 'Edytuj',
 	'communitypage-page-list-create' => 'Utwórz',
@@ -750,17 +838,17 @@ $messages['pl'] = array(
 	'communitypage-tasks-title' => 'Zadania',
 	'communitypage-entry-tasks' => 'Pomóż nam edytować tę wikię',
 	'communitypage-tasks-header-welcome' => 'Pomóż nam rozwijać {{SITENAME}}!',
-	'communitypage-tasks-admin-welcome' => 'Witaj na {{SITENAME}}. Ta wikia potrzebuje Twojego wkładu i pomocy. Oto przykładowe strony, którym dobrze zrobią twoje edycje. Przyjemnego tworzenia!',
+	'communitypage-tasks-admin-welcome' => 'Witaj na {{SITENAME}}. Ta wikia potrzebuje twojego wkładu i pomocy. Oto przykładowe strony, którym dobrze zrobią twoje edycje. Przyjemnego tworzenia!',
 	'communitypage-invite-friends' => 'Zaproś znajomych',
 	'communitypage-this-month' => 'Obecny miesiąc',
-	'communitypage-pages' => 'Stron',
+	'communitypage-pages' => 'Strony',
 	'communitypage-pageviews' => 'Odsłony',
 	'communitypage-edits' => 'Edycje',
 	'communitypage-editors' => 'Współtwórców',
-	'communitypage-top-contributors-week' => 'Najlepsi współtwórcy tygodnia',
+	'communitypage-top-contributors-week' => 'Najaktywniejsi współtwórcy tygodnia',
 	'communitypage-user-rank' => 'Twoja pozycja',
-	'communitypage-user-contributions' => '{{PLURAL:$1|edycja|edycji}}',
-	'communitypage-contributions' => '$1 {{PLURAL:$1|edycja|edycji}}',
+	'communitypage-user-contributions' => '{{PLURAL:$1|edycja|edycje|edycji}}',
+	'communitypage-contributions' => '$1 {{PLURAL:$1|edycja|edycje|edycji}}',
 	'communitypage-admins' => 'Administratorzy',
 	'communitypage-other-admins' => 'Inni administratorzy',
 	'communitypage-recently-joined' => 'Ostatni nowi współtwórcy',
@@ -772,28 +860,28 @@ $messages['pl'] = array(
 	'communitypage-edited' => 'edytował',
 	'communitypage-deleted' => 'usunął',
 	'communitypage-admin' => 'Administrator',
-	'communitypage-joined' => 'Dołączył do',
+	'communitypage-joined' => 'Dołączył/-a do',
 	'communitypage-modal-title' => 'Wszyscy współtwórcy',
 	'communitypage-modal-tab-all' => 'Wszyscy',
 	'communitypage-modal-tab-admins' => 'Administratorzy',
 	'communitypage-modal-tab-loadingerror' => 'Błąd przesyłania danych',
-	'communitypage-no-contributions' => 'Brak edycji w tym tygodniu! Bądź pierwszy!',
-	'communitypage-no-admins' => 'Ta wikia nie ma jeszcze Administratorów.',
-	'communitypage-no-admins-contact' => 'Skontaktuj się z nami, aby być pierwszy!',
+	'communitypage-no-contributions' => 'Brak edycji w tym tygodniu! Bądź {{GENDER:|pierwszy|pierwsza}}!',
+	'communitypage-no-admins' => 'Ta wikia nie ma jeszcze administratorów.',
+	'communitypage-no-admins-contact' => 'Skontaktuj się z nami, aby być {{GENDER:|pierwszy|pierwsza}}!',
 	'communitypage-communitycentral-link' => 'http://community.wikia.com/wiki/Special:Contact/adoption',
 	'communitypage-no-members' => 'Ta społeczność nie ma członków',
-	'communitypage-help-us-grow' => "Pomóż nam rozwijać '''{{SITENAME}} '''!",
+	'communitypage-help-us-grow' => "Pomóż nam rozwijać '''{{SITENAME}}'''!",
 	'communitypage-new' => 'Nowość',
 	'communitypage-popularpages-title' => 'Popularne strony',
 	'communitypage-popularpages-description' => 'Te strony zyskują na popularności wśród czytelników! Zadbaj, aby spełniały twoje wymagania oraz by zawierały aktualne obrazy i informacje.',
 	'communitypage-uncategorizedpages-title' => 'Strony bez kategorii',
-	'communitypage-uncategorizedpages-description' => 'Dodaj kategorie do stron. Poprawne określenie kategorie pomoże Ci w organizacji wikii!',
+	'communitypage-uncategorizedpages-description' => 'Dodaj kategorie do tych stron. Poprawnie określone kategorie są kluczowe dla dobrze zorganizowanej wikii!',
 	'communitypage-wantedpages-title' => 'Potrzebne strony',
 	'communitypage-wantedpages-description' => 'Istnieją redlinki kierujące do tych stron, ale te strony nie istnieją! Stwórz artykuł, albo przekieruj do właściwego artykułu, aby pomóc czytelnikom poruszać się po twojej społeczności.',
 	'communitypage-lastrevision' => 'Ostatnio edytowane przez $1, $2',
 	'communitypage-noofviews' => '$1 odsłon',
 	'communitypage-full-list' => 'Zobacz pełną listę',
-	'communitypage-modal-tab-all-contribution-header' => 'Ostatnia zmiana',
+	'communitypage-modal-tab-all-contribution-header' => 'Ostatnia edycja',
 	'communitypage-deadendpages-title' => 'Strony bez odnośników',
 	'communitypage-deadendpages-description' => 'Te strony powinny zawierać odnośniki do innych artykułów na twojej wikii. Dobrze powiązane artykuły pomagają czytelnikom odnaleźć więcej informacji i poprawiają pozycje w wyszukiwarkach.',
 	'communitypage-view-more' => 'Zobacz więcej',
@@ -808,22 +896,32 @@ $messages['pl'] = array(
 	'communitypage-policy-module-link-text' => 'Zobacz politykę',
 	'communitypage-entrypoint-modal-title' => 'Społeczność $1 jest tworzona przez $2 fanów takich jak ty.',
 	'communitypage-entrypoint-modal-edit-title' => 'Edytuj',
-	'communitypage-entrypoint-modal-edit-text' => 'Pomagaj innym użytkownikom pisać artykuly, organizować informacje oraz przesyłać zdjęcia, aby stworzyć najbardziej niesamowitą społeczność $1.',
+	'communitypage-entrypoint-modal-edit-text' => 'Pomagaj innym użytkownikom pisać artykuły, organizować informacje oraz przesyłać zdjęcia, aby stworzyć najbardziej niesamowitą społeczność $1.',
 	'communitypage-entrypoint-modal-connect-title' => 'Połącz się z innymi',
 	'communitypage-entrypoint-modal-connect-text' => 'Rozmawiaj na temat swoich reakcji, pomysłów i ulubionych tematów z innymi fanami, którzy kochają $1, tak jak ty.',
 	'communitypage-entrypoint-modal-explore-title' => 'Przeglądaj',
-	'communitypage-entrypoint-modal-explore-text' => 'Badź częścią czegoś wielkiego. Społeczność $1 jest tylko jedną z 360.000 społeczności w sieci Wikia.',
+	'communitypage-entrypoint-modal-explore-text' => 'Bądź częścią czegoś wielkiego. Społeczność $1 jest tylko jedną z 360&nbsp;000 społeczności w sieci Wikia.',
 	'communitypage-entrypoint-modal-button-text' => 'Pomóż',
 	'communitypage-entrypoint-modal-video-button-text' => 'Obejrzyj film',
 	'communitypage-anon-contrib-header' => '$1 lub $2 i dokonaj edycji tej wiki, aby zobaczyć jakie miejsce zajmujesz w rankingu tej społeczności!',
 	'communitypage-anon-login' => 'Zaloguj się',
 	'communitypage-anon-register' => 'zarejestruj się',
 	'communitypage-todo-module-page-name' => 'Lista-rzeczy-do-zrobienia-Społeczności',
-	'communitypage-todo-module-heading' => 'Lista-rzeczy-do-zrobienia-Społeczności',
+	'communitypage-todo-module-heading' => 'Lista rzeczy do zrobienia',
 	'communitypage-todo-module-edit-list' => 'Edytuj listę',
-	'communitypage-todo-module-description' => 'This is an area for admins to suggest pages that need work to help new contributors get started. (Admins - remember to provide details on what needs to be done.)',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-description' => 'Jest to obszar dla administratorów, w którym mogą sugerować strony, które wymagają pracy, aby pomóc nowym użytkownikom zacząć pracę. (Administratorzy - Pamiętajcie, aby podać szczegóły tego, co powinno być zrobione)',
+	'communitypage-todo-module-zero-state' => 'Chcesz dowiedzieć się jak możesz pomóc tej społeczności? Skontaktuj się z administratorem i poproś, żeby dodał tutaj sugestie.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Skontaktuj się z administratorem',
+	'communitypage-help-module-edit-page-name' => 'Edytowanie',
+	'communitypage-help-module-add-link-name' => 'Linki',
+	'communitypage-help-module-new-page-name' => 'Nowa_strona',
+	'communitypage-help-edit-page' => 'Jak edytować stronę',
+	'communitypage-help-add-link' => 'Jak dodać link',
+	'communitypage-help-add-new-page' => 'Jak dodać nową stronę',
+	'communitypage-first-edit-heading' => 'Gratulacje z okazji wykonania pierwszej edycji na tej wikii!',
+	'communitypage-first-edit-subheading' => 'Odwiedź stronę społeczności, aby poznać innych i znaleźć więcej sposobów jak pomóc',
+	'communitypage-first-edit-get-started' => 'Dalej',
+	'communitypage-first-edit-maybe-later' => 'Może później',
 );
 
 $messages['pt'] = array(
@@ -915,8 +1013,18 @@ $messages['pt'] = array(
 	'communitypage-todo-module-heading' => 'Lista de tarefas da Comunidade',
 	'communitypage-todo-module-edit-list' => 'Editar lista',
 	'communitypage-todo-module-description' => 'Esta é uma área para administradores sugerirem páginas que precisam de atenção para ajudar aos novos contribuidores a começar. (Administradores - lembrem-se de inserir detalhes sobre o que precisa ser feito).',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => 'Quer saber como você pode ajudar essa comunidade? Fale com um administrador e peça sugestões.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Fale com um administrador',
+	'communitypage-help-module-edit-page-name' => 'Editando',
+	'communitypage-help-module-add-link-name' => 'Links',
+	'communitypage-help-module-new-page-name' => 'Nova_página',
+	'communitypage-help-edit-page' => 'Como editar uma página',
+	'communitypage-help-add-link' => 'Como adicionar um link',
+	'communitypage-help-add-new-page' => 'Como adicionar uma nova página',
+	'communitypage-first-edit-heading' => 'Parabéns por fazer a sua primeira edição nesta comunidade!',
+	'communitypage-first-edit-subheading' => 'Visite a Página da Comunidade para conhecer outras pessoas e encontrar mais maneiras de ajudar',
+	'communitypage-first-edit-get-started' => 'Continuar',
+	'communitypage-first-edit-maybe-later' => 'Talvez mais tarde',
 );
 
 $messages['ru'] = array(
@@ -1008,8 +1116,18 @@ $messages['ru'] = array(
 	'communitypage-todo-module-heading' => 'Список задач сообщества',
 	'communitypage-todo-module-edit-list' => 'Править список',
 	'communitypage-todo-module-description' => 'Здесь администраторы размещают информацию о страницах, нуждающихся в доработке, чтобы новые участники могли сразу же приступить к правке. (Администраторы, не забудьте подробно объяснить, что необходимо сделать.)',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => 'Хотите узнать, как вы можете помочь этому сообществу? $1и попросите сообщить о страницах, нуждающихся в доработке, здесь.',
 	'communitypage-todo-module-zero-state-contact-admin' => 'Свяжитесь с администратором',
+	'communitypage-help-module-edit-page-name' => 'Редактирование',
+	'communitypage-help-module-add-link-name' => 'Ссылки',
+	'communitypage-help-module-new-page-name' => 'New_page',
+	'communitypage-help-edit-page' => 'Редактирование страницы',
+	'communitypage-help-add-link' => 'Добавление ссылки',
+	'communitypage-help-add-new-page' => 'Добавление новой статьи',
+	'communitypage-first-edit-heading' => 'Поздравляем с первой правкой на этой вики!',
+	'communitypage-first-edit-subheading' => 'Познакомьтесь с другими участниками и узнайте, как помочь вики, на Странице сообщества',
+	'communitypage-first-edit-get-started' => 'Приступить к работе',
+	'communitypage-first-edit-maybe-later' => 'Возможно, позже',
 );
 
 $messages['zh-hans'] = array(
@@ -1101,8 +1219,18 @@ $messages['zh-hans'] = array(
 	'communitypage-todo-module-heading' => '社区待办事项列表',
 	'communitypage-todo-module-edit-list' => '编辑列表',
 	'communitypage-todo-module-description' => '管理员可以在这里建议新的贡献者通过完成哪些不完善的页面帮助社区发展。（管理员请注意：您需要提供详细说明）。',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => '想要知道如何对这个社区进行贡献吗？联系管理员并且告诉他们在这里提供建议。',
 	'communitypage-todo-module-zero-state-contact-admin' => '联系管理员',
+	'communitypage-help-module-edit-page-name' => '编辑中',
+	'communitypage-help-module-add-link-name' => '链接',
+	'communitypage-help-module-new-page-name' => '新页面',
+	'communitypage-help-edit-page' => '如何编辑页面',
+	'communitypage-help-add-link' => '如何添加链接',
+	'communitypage-help-add-new-page' => '如何添加新的页面',
+	'communitypage-first-edit-heading' => '恭喜你对这个社区做出了首次编辑！',
+	'communitypage-first-edit-subheading' => '访问社区页认识其他朋友，了解如何更好地为社区做出贡献！ ',
+	'communitypage-first-edit-get-started' => '立即开始',
+	'communitypage-first-edit-maybe-later' => '以后再说',
 );
 
 $messages['zh-hant'] = array(
@@ -1194,7 +1322,17 @@ $messages['zh-hant'] = array(
 	'communitypage-todo-module-heading' => '社區待辦事項清單',
 	'communitypage-todo-module-edit-list' => '編輯清單',
 	'communitypage-todo-module-description' => '管理員可以在這裡建議新的貢獻者通過完成哪些不完善的頁面幫助社區發展。（管理員請注意：您需要提供詳細說明）。',
-	'communitypage-todo-module-zero-state' => 'Want to know how you can contribute to this community? Contact an admin and ask them to provide suggestions here.',
+	'communitypage-todo-module-zero-state' => '想要知道你可以如何對這個社區進行貢獻嗎？聯繫管理員並且告訴他們在這裡提供建議。',
 	'communitypage-todo-module-zero-state-contact-admin' => '聯繫管理員',
+	'communitypage-help-module-edit-page-name' => '編輯中',
+	'communitypage-help-module-add-link-name' => '連結',
+	'communitypage-help-module-new-page-name' => '新頁面',
+	'communitypage-help-edit-page' => '如何編輯頁面',
+	'communitypage-help-add-link' => '如何添加連結',
+	'communitypage-help-add-new-page' => '如何添加新的頁面',
+	'communitypage-first-edit-heading' => '恭喜你對這個社區做出了首次編輯！',
+	'communitypage-first-edit-subheading' => '訪問社區頁認識其他朋友，瞭解如何更好地為社區做出貢獻！',
+	'communitypage-first-edit-get-started' => '現在開始',
+	'communitypage-first-edit-maybe-later' => '以後再說',
 );
 
