@@ -1,6 +1,7 @@
 <?php
 
 class CommunityPageSpecialHooks {
+	const FIRST_EDIT_COOKIE_KEY = 'community-page-first-time';
 
 	/**
 	 * Cache key invalidation when an article is edited
@@ -62,7 +63,9 @@ class CommunityPageSpecialHooks {
 	 * @return true
 	 */
 	public static function onBeforePageDisplay( \OutputPage $out, \Skin $skin ) {
-		if ( $out->getUser()->isAnon() &&
+		$user = $out->getUser();
+
+		if ( $user->isAnon() &&
 			!isset( $_COOKIE['cpBenefitsModalShown'] ) &&
 			$out->getRequest()->getVal( 'action' ) !== 'edit' &&
 			$out->getRequest()->getVal( 'veaction' ) !== 'edit' &&
@@ -71,6 +74,12 @@ class CommunityPageSpecialHooks {
 			\Wikia::addAssetsToOutput( 'community_page_benefits_js' );
 			\Wikia::addAssetsToOutput( 'community_page_benefits_scss' );
 		}
+
+		if ( !$user->isAnon() && !$user->isAllowed( 'first-edit-dialog-exempt' ) ) {
+			\Wikia::addAssetsToOutput( 'community_page_new_user_modal_js' );
+			\Wikia::addAssetsToOutput( 'community_page_new_user_modal_scss' );
+		}
+
 		return true;
 	}
 
@@ -108,11 +117,20 @@ class CommunityPageSpecialHooks {
 	}
 
 	public static function onUserFirstEditOnLocalWiki( $userId, $wikiId ) {
+		global $wgCookieDomain, $wgCookiePath;
+
 		$key = CommunityPageSpecialUsersModel::getMemcKey(
 			[ CommunityPageSpecialUsersModel::RECENTLY_JOINED_MCACHE_KEY, 14 ]
 		);
 		WikiaDataAccess::cachePurge( $key );
 		CommunityPageSpecialUsersModel::logUserModelPerformanceData( 'purge', 'recently_joined' );
+
+		// Set cookie to show first edit modal to user
+		$user = User::newFromId( $userId );
+
+		if ( !$user->isAllowed( 'first-edit-dialog-exempt' ) ) {
+			setcookie( self::FIRST_EDIT_COOKIE_KEY, true, time()+60, $wgCookiePath, $wgCookieDomain );
+		}
 
 		return true;
 	}
