@@ -109,7 +109,7 @@ require([
 				controller: 'CommunityPageSpecial',
 				method: tab.request,
 				data: {
-					uselang: window.wgUserLanguage
+					uselang: mw.config.get('wgUserLanguage')
 				},
 				format: 'json',
 				type: 'get',
@@ -250,6 +250,18 @@ require([
 			if (category !== null && category.length > 0) {
 				handleClick(event, category);
 			}
+		});
+
+		// Track clicks in the To-do List module
+		$('.community-page-todo-list-module-edit').on('mousedown touchstart', 'a', function (event) {
+			handleClick(event, 'community-page-todo-list-module');
+		});
+
+		$('.community-page-todo-list-module-content').on('mousedown touchstart', 'a', function (event) {
+			track({
+				category: 'community-page-todo-list-module',
+				label: 'community-page-todo-list-module-content',
+			});
 		});
 	}
 
