@@ -43,7 +43,7 @@ describe('ext.createNewWiki.helper', function () {
 				wikiDomain: 'foo',
 				wikiLanguage: 'pl',
 				toStringOutput: '?wikiName=foo&wikiDomain=foo&wikiLanguage=pl',
-				expected: '/signin?redirect=%3FwikiName%3Dfoo%26wikiDomain%3Dfoo%26wikiLanguage%3Dpl'
+				expected: '/register?redirect=%3FwikiName%3Dfoo%26wikiDomain%3Dfoo%26wikiLanguage%3Dpl'
 			}
 		];
 
