@@ -46,7 +46,7 @@ class UserTest extends WikiaBaseTest {
 		$this->userPreferenceServiceMock = $this->getMock( PreferenceService::class,
 			['getGlobalPreference', 'getPreferences', 'setPreferences', 'setGlobalPreference', 'deleteGlobalPreference',
 			'getLocalPreference', 'setLocalPreference', 'deleteLocalPreference', 'save', 'getGlobalDefault', 'deleteFromCache',
-			'deleteAllPreferences', 'findWikisWithLocalPreferenceValue'] );
+			'deleteAllPreferences', 'findWikisWithLocalPreferenceValue', 'findUsersWithGlobalPreferenceValue'] );
 
 		$this->userAttributeServiceMock = $this->getMock( AttributeService::class );
 		$this->userAttributesMock = $this->getMockBuilder( UserAttributes::class )
