@@ -10,21 +10,25 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -55,21 +59,24 @@ class ApiClient
 
     /**
      * Configuration
+     *
      * @var Configuration
      */
     protected $config;
 
     /**
      * Object Serializer
+     *
      * @var ObjectSerializer
      */
     protected $serializer;
 
     /**
      * Constructor of the class
+     *
      * @param Configuration $config config for this ApiClient
      */
-    public function __construct(Configuration $config = null)
+    public function __construct(\Swagger\Client\Configuration $config = null)
     {
         if ($config == null) {
             $config = Configuration::getDefaultConfiguration();
@@ -81,6 +88,7 @@ class ApiClient
 
     /**
      * Get the config
+     *
      * @return Configuration
      */
     public function getConfig()
@@ -90,6 +98,7 @@ class ApiClient
 
     /**
      * Get the serializer
+     *
      * @return ObjectSerializer
      */
     public function getSerializer()
@@ -99,7 +108,9 @@ class ApiClient
 
     /**
      * Get API key (with prefix if set)
+     *
      * @param  string $apiKeyIdentifier name of apikey
+     *
      * @return string API key with the prefix
      */
     public function getApiKeyWithPrefix($apiKeyIdentifier)
@@ -122,16 +133,19 @@ class ApiClient
 
     /**
      * Make the HTTP call (Sync)
+     *
      * @param string $resourcePath path to method endpoint
      * @param string $method       method to call
      * @param array  $queryParams  parameters to be place in query URL
      * @param array  $postData     parameters to be placed in POST body
      * @param array  $headerParams parameters to be place in request header
      * @param string $responseType expected response type of the endpoint
+     * @param string $endpointPath path to method endpoint before expanding parameters
+     *
      * @throws \Swagger\Client\ApiException on a non 2xx response
      * @return mixed
      */
-    public function callApi($resourcePath, $method, $queryParams, $postData, $headerParams, $responseType = null)
+    public function callApi($resourcePath, $method, $queryParams, $postData, $headerParams, $responseType = null, $endpointPath = null)
     {
 
         $headers = array();
@@ -160,7 +174,7 @@ class ApiClient
         if ($this->config->getCurlTimeout() != 0) {
             curl_setopt($curl, CURLOPT_TIMEOUT, $this->config->getCurlTimeout());
         }
-        // return the result on success, rather than just true 
+        // return the result on success, rather than just true
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
 
         curl_setopt($curl, CURLOPT_HTTPHEADER, $headers);
@@ -171,7 +185,7 @@ class ApiClient
             curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
         }
 
-        if (! empty($queryParams)) {
+        if (!empty($queryParams)) {
             $url = ($url . '?' . http_build_query($queryParams));
         }
 
@@ -202,7 +216,7 @@ class ApiClient
 
         // debugging for curl
         if ($this->config->getDebug()) {
-            error_log("[DEBUG] HTTP Request body  ~BEGIN~\n".print_r($postData, true)."\n~END~\n", 3, $this->config->getDebugFile());
+            error_log("[DEBUG] HTTP Request body  ~BEGIN~".PHP_EOL.print_r($postData, true).PHP_EOL."~END~".PHP_EOL, 3, $this->config->getDebugFile());
 
             curl_setopt($curl, CURLOPT_VERBOSE, 1);
             curl_setopt($curl, CURLOPT_STDERR, fopen($this->config->getDebugFile(), 'a'));
@@ -216,19 +230,19 @@ class ApiClient
         // Make the request
         $response = curl_exec($curl);
         $http_header_size = curl_getinfo($curl, CURLINFO_HEADER_SIZE);
-        $http_header = $this->http_parse_headers(substr($response, 0, $http_header_size));
+        $http_header = $this->httpParseHeaders(substr($response, 0, $http_header_size));
         $http_body = substr($response, $http_header_size);
         $response_info = curl_getinfo($curl);
 
         // debug HTTP response body
         if ($this->config->getDebug()) {
-            error_log("[DEBUG] HTTP Response body ~BEGIN~\n".print_r($http_body, true)."\n~END~\n", 3, $this->config->getDebugFile());
+            error_log("[DEBUG] HTTP Response body ~BEGIN~".PHP_EOL.print_r($http_body, true).PHP_EOL."~END~".PHP_EOL, 3, $this->config->getDebugFile());
         }
 
         // Handle the response
         if ($response_info['http_code'] == 0) {
             throw new ApiException("API call to $url timed out: ".serialize($response_info), 0, null, null);
-        } elseif ($response_info['http_code'] >= 200 && $response_info['http_code'] <= 299 ) {
+        } elseif ($response_info['http_code'] >= 200 && $response_info['http_code'] <= 299) {
             // return raw body if response is a file
             if ($responseType == '\SplFileObject' || $responseType == 'string') {
                 return array($http_body, $response_info['http_code'], $http_header);
@@ -246,7 +260,9 @@ class ApiClient
 
             throw new ApiException(
                 "[".$response_info['http_code']."] Error connecting to the API ($url)",
-                $response_info['http_code'], $http_header, $data
+                $response_info['http_code'],
+                $http_header,
+                $data
             );
         }
         return array($data, $response_info['http_code'], $http_header);
@@ -259,7 +275,7 @@ class ApiClient
      *
      * @return string Accept (e.g. application/json)
      */
-    public static function selectHeaderAccept($accept)
+    static public function selectHeaderAccept($accept)
     {
         if (count($accept) === 0 or (count($accept) === 1 and $accept[0] === '')) {
             return null;
@@ -277,7 +293,7 @@ class ApiClient
      *
      * @return string Content-Type (e.g. application/json)
      */
-    public static function selectHeaderContentType($content_type)
+    static public function selectHeaderContentType($content_type)
     {
         if (count($content_type) === 0 or (count($content_type) === 1 and $content_type[0] === '')) {
             return 'application/json';
@@ -295,44 +311,35 @@ class ApiClient
     *
     * @return string[] Array of HTTP response heaers
     */
-    protected function http_parse_headers($raw_headers)
+    protected function httpParseHeaders($raw_headers)
     {
         // ref/credit: http://php.net/manual/en/function.http-parse-headers.php#112986
         $headers = array();
-        $key = ''; // [+]
-   
-        foreach(explode("\n", $raw_headers) as $i => $h)
-        {
+        $key = '';
+
+        foreach (explode("\n", $raw_headers) as $h) {
             $h = explode(':', $h, 2);
-   
-            if (isset($h[1]))
-            {
-                if (!isset($headers[$h[0]]))
+
+            if (isset($h[1])) {
+                if (!isset($headers[$h[0]])) {
                     $headers[$h[0]] = trim($h[1]);
-                elseif (is_array($headers[$h[0]]))
-                {
-                    // $tmp = array_merge($headers[$h[0]], array(trim($h[1]))); // [-]
-                    // $headers[$h[0]] = $tmp; // [-]
-                    $headers[$h[0]] = array_merge($headers[$h[0]], array(trim($h[1]))); // [+]
+                } elseif (is_array($headers[$h[0]])) {
+                    $headers[$h[0]] = array_merge($headers[$h[0]], array(trim($h[1])));
+                } else {
+                    $headers[$h[0]] = array_merge(array($headers[$h[0]]), array(trim($h[1])));
                 }
-                else
-                {
-                    // $tmp = array_merge(array($headers[$h[0]]), array(trim($h[1]))); // [-]
-                    // $headers[$h[0]] = $tmp; // [-]
-                    $headers[$h[0]] = array_merge(array($headers[$h[0]]), array(trim($h[1]))); // [+]
+
+                $key = $h[0];
+            } else {
+                if (substr($h[0], 0, 1) == "\t") {
+                    $headers[$key] .= "\r\n\t".trim($h[0]);
+                } elseif (!$key) {
+                    $headers[0] = trim($h[0]);
                 }
-   
-                $key = $h[0]; // [+]
+                trim($h[0]);
             }
-            else // [+]
-            { // [+]
-                if (substr($h[0], 0, 1) == "\t") // [+]
-                    $headers[$key] .= "\r\n\t".trim($h[0]); // [+]
-                elseif (!$key) // [+]
-                    $headers[0] = trim($h[0]);trim($h[0]); // [+]
-            } // [+]
         }
-   
+
         return $headers;
     }
 }
