@@ -9,21 +9,25 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -48,30 +52,34 @@ class ApiException extends Exception
 
     /**
      * The HTTP body of the server response either as Json or string.
+     *
      * @var mixed
      */
     protected $responseBody;
 
     /**
      * The HTTP header of the server response.
+     *
      * @var string[]
      */
     protected $responseHeaders;
 
     /**
      * The deserialized response object
+     *
      * @var $responseObject;
      */
     protected $responseObject;
 
     /**
      * Constructor
+     *
      * @param string $message         Error message
      * @param int    $code            HTTP status code
      * @param string $responseHeaders HTTP response header
      * @param mixed  $responseBody    HTTP body of the server response either as Json or string
      */
-    public function __construct($message="", $code=0, $responseHeaders=null, $responseBody=null)
+    public function __construct($message = "", $code = 0, $responseHeaders = null, $responseBody = null)
     {
         parent::__construct($message, $code);
         $this->responseHeaders = $responseHeaders;
@@ -100,7 +108,9 @@ class ApiException extends Exception
 
     /**
      * Sets the deseralized response object (during deserialization)
+     *
      * @param mixed $obj Deserialized response object
+     *
      * @return void
      */
     public function setResponseObject($obj)
