@@ -11,19 +11,21 @@
  */
 
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 /**
@@ -47,7 +49,7 @@ namespace Swagger\Client;
 class Configuration
 {
 
-    private static $_defaultConfiguration = null;
+    private static $defaultConfiguration = null;
 
     /**
      * Associate array to store API key(s)
@@ -96,7 +98,7 @@ class Configuration
      *
      * @var string
      */
-    protected $host = 'https://localhost/';
+    protected $host = 'https://localhost/user-preference';
 
     /**
      * Timeout (second) of the HTTP request, by default set to 0, no timeout
@@ -487,11 +489,11 @@ class Configuration
      */
     public static function getDefaultConfiguration()
     {
-        if (self::$_defaultConfiguration == null) {
-            self::$_defaultConfiguration = new Configuration();
+        if (self::$defaultConfiguration == null) {
+            self::$defaultConfiguration = new Configuration();
         }
 
-        return self::$_defaultConfiguration;
+        return self::$defaultConfiguration;
     }
 
     /**
@@ -503,7 +505,7 @@ class Configuration
      */
     public static function setDefaultConfiguration(Configuration $config)
     {
-        self::$_defaultConfiguration = $config;
+        self::$defaultConfiguration = $config;
     }
 
     /**
@@ -513,14 +515,12 @@ class Configuration
      */
     public static function toDebugReport()
     {
-        $report  = "PHP SDK (Swagger\Client) Debug Report:\n";
-        $report .= "    OS: ".php_uname()."\n";
-        $report .= "    PHP Version: ".phpversion()."\n";
-        $report .= "    OpenAPI Spec Version: \n";
-        $report .= "    SDK Package Version: 1.0.0\n";
-        $report .= "    Temp Folder Path: ".self::getDefaultConfiguration()->getTempFolderPath()."\n";
+        $report  = 'PHP SDK (Swagger\Client) Debug Report:' . PHP_EOL;
+        $report .= '    OS: ' . php_uname() . PHP_EOL;
+        $report .= '    PHP Version: ' . phpversion() . PHP_EOL;
+        $report .= '    OpenAPI Spec Version: ' . PHP_EOL;
+        $report .= '    Temp Folder Path: ' . self::getDefaultConfiguration()->getTempFolderPath() . PHP_EOL;
 
         return $report;
     }
-
 }
