@@ -1,6 +1,6 @@
 <?php
 /**
- * ObjectSerializer 
+ * ObjectSerializer
  *
  * PHP version 5
  *
@@ -10,21 +10,25 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -37,7 +41,7 @@ namespace Swagger\Client;
  * ObjectSerializer Class Doc Comment
  *
  * @category Class
- * @package  Swagger\Client 
+ * @package  Swagger\Client
  * @author   http://github.com/swagger-api/swagger-codegen
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
@@ -55,14 +59,14 @@ class ObjectSerializer
     public static function sanitizeForSerialization($data)
     {
         if (is_scalar($data) || null === $data) {
-            $sanitized = $data;
+            return $data;
         } elseif ($data instanceof \DateTime) {
-            $sanitized = $data->format(\DateTime::ATOM);
+            return $data->format(\DateTime::ATOM);
         } elseif (is_array($data)) {
             foreach ($data as $property => $value) {
                 $data[$property] = self::sanitizeForSerialization($value);
             }
-            $sanitized = $data;
+            return $data;
         } elseif (is_object($data)) {
             $values = array();
             foreach (array_keys($data::swaggerTypes()) as $property) {
@@ -71,17 +75,15 @@ class ObjectSerializer
                     $values[$data::attributeMap()[$property]] = self::sanitizeForSerialization($data->$getter());
                 }
             }
-            $sanitized = (object)$values;
+            return (object)$values;
         } else {
-            $sanitized = (string)$data;
+            return (string)$data;
         }
-
-        return $sanitized;
     }
 
     /**
      * Sanitize filename by removing path.
-     * e.g. ../../sun.gif becomes sun.gif 
+     * e.g. ../../sun.gif becomes sun.gif
      *
      * @param string $filename filename to be sanitized
      *
@@ -187,7 +189,7 @@ class ObjectSerializer
      *
      * @return string
      */
-    public function serializeCollection(array $collection, $collectionFormat, $allowCollectionFormatMulti=false)
+    public function serializeCollection(array $collection, $collectionFormat, $allowCollectionFormatMulti = false)
     {
         if ($allowCollectionFormatMulti && ('multi' === $collectionFormat)) {
             // http_build_query() almost does the job for us. We just
@@ -221,10 +223,10 @@ class ObjectSerializer
      *
      * @return object an instance of $class
      */
-    public static function deserialize($data, $class, $httpHeaders=null, $discriminator=null)
+    public static function deserialize($data, $class, $httpHeaders = null, $discriminator = null)
     {
         if (null === $data) {
-            $deserialized = null;
+            return null;
         } elseif (substr($class, 0, 4) === 'map[') { // for associative array e.g. map[string,int]
             $inner = substr($class, 4, -1);
             $deserialized = array();
@@ -235,16 +237,17 @@ class ObjectSerializer
                     $deserialized[$key] = self::deserialize($value, $subClass, null, $discriminator);
                 }
             }
+            return $deserialized;
         } elseif (strcasecmp(substr($class, -2), '[]') == 0) {
             $subClass = substr($class, 0, -2);
             $values = array();
             foreach ($data as $key => $value) {
                 $values[] = self::deserialize($value, $subClass, null, $discriminator);
             }
-            $deserialized = $values;
+            return $values;
         } elseif ($class === 'object') {
             settype($data, 'array');
-            $deserialized = $data;
+            return $data;
         } elseif ($class === '\DateTime') {
             // Some API's return an invalid, empty string as a
             // date-time property. DateTime::__construct() will return
@@ -253,24 +256,29 @@ class ObjectSerializer
             // be interpreted as a missing field/value. Let's handle
             // this graceful.
             if (!empty($data)) {
-                $deserialized = new \DateTime($data);
+                return new \DateTime($data);
             } else {
-                $deserialized = null;
+                return null;
             }
         } elseif (in_array($class, array('void', 'bool', 'string', 'double', 'byte', 'mixed', 'integer', 'float', 'int', 'DateTime', 'number', 'boolean', 'object'))) {
             settype($data, $class);
-            $deserialized = $data;
+            return $data;
         } elseif ($class === '\SplFileObject') {
             // determine file name
-            if (array_key_exists('Content-Disposition', $httpHeaders) && preg_match('/inline; filename=[\'"]?([^\'"\s]+)[\'"]?$/i', $httpHeaders['Content-Disposition'], $match)) {
+            if (array_key_exists('Content-Disposition', $httpHeaders) &&
+                preg_match('/inline; filename=[\'"]?([^\'"\s]+)[\'"]?$/i', $httpHeaders['Content-Disposition'], $match)) {
                 $filename = Configuration::getDefaultConfiguration()->getTempFolderPath() . sanitizeFilename($match[1]);
             } else {
                 $filename = tempnam(Configuration::getDefaultConfiguration()->getTempFolderPath(), '');
             }
             $deserialized = new \SplFileObject($filename, "w");
             $byte_written = $deserialized->fwrite($data);
-            error_log("[INFO] Written $byte_written byte to $filename. Please move the file to a proper folder or delete the temp file after processing.\n", 3, Configuration::getDefaultConfiguration()->getDebugFile());
-      
+ 
+            if (Configuration::getDefaultConfiguration()->getDebug()) {
+                error_log("[DEBUG] Written $byte_written byte to $filename. Please move the file to a proper folder or delete the temp file after processing.".PHP_EOL, 3, Configuration::getDefaultConfiguration()->getDebugFile());
+            }
+
+            return $deserialized;
         } else {
             // If a discriminator is defined and points to a valid subclass, use it.
             if (!empty($discriminator) && isset($data->{$discriminator}) && is_string($data->{$discriminator})) {
@@ -282,19 +290,17 @@ class ObjectSerializer
             $instance = new $class();
             foreach ($instance::swaggerTypes() as $property => $type) {
                 $propertySetter = $instance::setters()[$property];
-     
+
                 if (!isset($propertySetter) || !isset($data->{$instance::attributeMap()[$property]})) {
                     continue;
                 }
-     
+
                 $propertyValue = $data->{$instance::attributeMap()[$property]};
                 if (isset($propertyValue)) {
                     $instance->$propertySetter(self::deserialize($propertyValue, $type, null, $discriminator));
                 }
             }
-            $deserialized = $instance;
+            return $instance;
         }
-     
-        return $deserialized;
     }
 }
