@@ -9,25 +9,28 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 /**
- * NOTE: This class is auto generated by the swagger code generator program. 
- * https://github.com/swagger-api/swagger-codegen 
+ * NOTE: This class is auto generated by the swagger code generator program.
+ * https://github.com/swagger-api/swagger-codegen
  * Do not edit the class manually.
  */
 
@@ -52,47 +55,51 @@ class ReverseLookupApi
 
     /**
      * API Client
+     *
      * @var \Swagger\Client\ApiClient instance of the ApiClient
      */
     protected $apiClient;
-  
+
     /**
      * Constructor
+     *
      * @param \Swagger\Client\ApiClient|null $apiClient The api client to use
      */
-    function __construct($apiClient = null)
+    public function __construct(\Swagger\Client\ApiClient $apiClient = null)
     {
         if ($apiClient == null) {
             $apiClient = new ApiClient();
             $apiClient->getConfig()->setHost('https://localhost/user-preference');
         }
-  
+
         $this->apiClient = $apiClient;
     }
-  
+
     /**
      * Get API client
+     *
      * @return \Swagger\Client\ApiClient get the API client
      */
     public function getApiClient()
     {
         return $this->apiClient;
     }
-  
+
     /**
      * Set the API client
+     *
      * @param \Swagger\Client\ApiClient $apiClient set the API client
+     *
      * @return ReverseLookupApi
      */
-    public function setApiClient(ApiClient $apiClient)
+    public function setApiClient(\Swagger\Client\ApiClient $apiClient)
     {
         $this->apiClient = $apiClient;
         return $this;
     }
-  
-    
+
     /**
-     * findUsersWithGlobalPreference
+     * Operation findUsersWithGlobalPreference
      *
      * Finds all users that has the global preference set to a value
      *
@@ -100,18 +107,17 @@ class ReverseLookupApi
      * @param string $value The preference value (optionally provided) (optional)
      * @param int $limit How many results to return (optional, default to 1000)
      * @param int $user_id_continue Find results after userId (results returned in ascending order) (optional, default to 0)
-     * @return void
+     * @return string[]
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function findUsersWithGlobalPreference($preference_name, $value = null, $limit = null, $user_id_continue = null)
     {
-        list($response, $statusCode, $httpHeader) = $this->findUsersWithGlobalPreferenceWithHttpInfo ($preference_name, $value, $limit, $user_id_continue);
-        return $response; 
+        list($response) = $this->findUsersWithGlobalPreferenceWithHttpInfo($preference_name, $value, $limit, $user_id_continue);
+        return $response;
     }
 
-
     /**
-     * findUsersWithGlobalPreferenceWithHttpInfo
+     * Operation findUsersWithGlobalPreferenceWithHttpInfo
      *
      * Finds all users that has the global preference set to a value
      *
@@ -119,45 +125,40 @@ class ReverseLookupApi
      * @param string $value The preference value (optionally provided) (optional)
      * @param int $limit How many results to return (optional, default to 1000)
      * @param int $user_id_continue Find results after userId (results returned in ascending order) (optional, default to 0)
-     * @return Array of null, HTTP status code, HTTP response headers (array of strings)
+     * @return Array of string[], HTTP status code, HTTP response headers (array of strings)
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function findUsersWithGlobalPreferenceWithHttpInfo($preference_name, $value = null, $limit = null, $user_id_continue = null)
     {
-        
         // verify the required parameter 'preference_name' is set
         if ($preference_name === null) {
             throw new \InvalidArgumentException('Missing the required parameter $preference_name when calling findUsersWithGlobalPreference');
         }
-  
         // parse inputs
         $resourcePath = "/reverse-lookup/global/{preferenceName}/users";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // query params
-        
         if ($value !== null) {
             $queryParams['value'] = $this->apiClient->getSerializer()->toQueryValue($value);
-        }// query params
-        
+        }
+        // query params
         if ($limit !== null) {
             $queryParams['limit'] = $this->apiClient->getSerializer()->toQueryValue($limit);
-        }// query params
-        
+        }
+        // query params
         if ($user_id_continue !== null) {
             $queryParams['userIdContinue'] = $this->apiClient->getSerializer()->toQueryValue($user_id_continue);
         }
-        
         // path params
-        
         if ($preference_name !== null) {
             $resourcePath = str_replace(
                 "{" . "preferenceName" . "}",
@@ -169,35 +170,39 @@ class ReverseLookupApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'GET',
-                $queryParams, $httpBody,
-                $headerParams
+                $resourcePath,
+                'GET',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                'string[]',
+                '/reverse-lookup/global/{preferenceName}/users'
             );
-            
-            return array(null, $statusCode, $httpHeader);
-            
+
+            return array($this->apiClient->getSerializer()->deserialize($response, 'string[]', $httpHeader), $statusCode, $httpHeader);
         } catch (ApiException $e) {
-            switch ($e->getCode()) { 
+            switch ($e->getCode()) {
+                case 200:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), 'string[]', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
-  
+
             throw $e;
         }
     }
-    
+
     /**
-     * findWikisWithLocalPreference
+     * Operation findWikisWithLocalPreference
      *
      * finds wikis where at least one user has a local preference set to a value
      *
@@ -208,13 +213,12 @@ class ReverseLookupApi
      */
     public function findWikisWithLocalPreference($preference_name, $value = null)
     {
-        list($response, $statusCode, $httpHeader) = $this->findWikisWithLocalPreferenceWithHttpInfo ($preference_name, $value);
-        return $response; 
+        list($response) = $this->findWikisWithLocalPreferenceWithHttpInfo($preference_name, $value);
+        return $response;
     }
 
-
     /**
-     * findWikisWithLocalPreferenceWithHttpInfo
+     * Operation findWikisWithLocalPreferenceWithHttpInfo
      *
      * finds wikis where at least one user has a local preference set to a value
      *
@@ -225,32 +229,27 @@ class ReverseLookupApi
      */
     public function findWikisWithLocalPreferenceWithHttpInfo($preference_name, $value = null)
     {
-        
         // verify the required parameter 'preference_name' is set
         if ($preference_name === null) {
             throw new \InvalidArgumentException('Missing the required parameter $preference_name when calling findWikisWithLocalPreference');
         }
-  
         // parse inputs
         $resourcePath = "/reverse-lookup/local/{preferenceName}/wikis";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // query params
-        
         if ($value !== null) {
             $queryParams['value'] = $this->apiClient->getSerializer()->toQueryValue($value);
         }
-        
         // path params
-        
         if ($preference_name !== null) {
             $resourcePath = str_replace(
                 "{" . "preferenceName" . "}",
@@ -262,39 +261,35 @@ class ReverseLookupApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'GET',
-                $queryParams, $httpBody,
-                $headerParams, 'string[]'
+                $resourcePath,
+                'GET',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                'string[]',
+                '/reverse-lookup/local/{preferenceName}/wikis'
             );
-            
-            if (!$response) {
-                return array(null, $statusCode, $httpHeader);
+
+            return array($this->apiClient->getSerializer()->deserialize($response, 'string[]', $httpHeader), $statusCode, $httpHeader);
+        } catch (ApiException $e) {
+            switch ($e->getCode()) {
+                case 200:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), 'string[]', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
 
-            return array(\Swagger\Client\ObjectSerializer::deserialize($response, 'string[]', $httpHeader), $statusCode, $httpHeader);
-            
-        } catch (ApiException $e) {
-            switch ($e->getCode()) { 
-            case 200:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), 'string[]', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            }
-  
             throw $e;
         }
     }
-    
+
 }
