@@ -9,25 +9,28 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 /**
- * NOTE: This class is auto generated by the swagger code generator program. 
- * https://github.com/swagger-api/swagger-codegen 
+ * NOTE: This class is auto generated by the swagger code generator program.
+ * https://github.com/swagger-api/swagger-codegen
  * Do not edit the class manually.
  */
 
@@ -52,47 +55,51 @@ class UserPreferencesApi
 
     /**
      * API Client
+     *
      * @var \Swagger\Client\ApiClient instance of the ApiClient
      */
     protected $apiClient;
-  
+
     /**
      * Constructor
+     *
      * @param \Swagger\Client\ApiClient|null $apiClient The api client to use
      */
-    function __construct($apiClient = null)
+    public function __construct(\Swagger\Client\ApiClient $apiClient = null)
     {
         if ($apiClient == null) {
             $apiClient = new ApiClient();
             $apiClient->getConfig()->setHost('https://localhost/user-preference');
         }
-  
+
         $this->apiClient = $apiClient;
     }
-  
+
     /**
      * Get API client
+     *
      * @return \Swagger\Client\ApiClient get the API client
      */
     public function getApiClient()
     {
         return $this->apiClient;
     }
-  
+
     /**
      * Set the API client
+     *
      * @param \Swagger\Client\ApiClient $apiClient set the API client
+     *
      * @return UserPreferencesApi
      */
-    public function setApiClient(ApiClient $apiClient)
+    public function setApiClient(\Swagger\Client\ApiClient $apiClient)
     {
         $this->apiClient = $apiClient;
         return $this;
     }
-  
-    
+
     /**
-     * deleteUserPreferences
+     * Operation deleteUserPreferences
      *
      * deletes all of a user's preferences
      *
@@ -102,13 +109,12 @@ class UserPreferencesApi
      */
     public function deleteUserPreferences($user_id)
     {
-        list($response, $statusCode, $httpHeader) = $this->deleteUserPreferencesWithHttpInfo ($user_id);
-        return $response; 
+        list($response) = $this->deleteUserPreferencesWithHttpInfo($user_id);
+        return $response;
     }
 
-
     /**
-     * deleteUserPreferencesWithHttpInfo
+     * Operation deleteUserPreferencesWithHttpInfo
      *
      * deletes all of a user's preferences
      *
@@ -118,28 +124,23 @@ class UserPreferencesApi
      */
     public function deleteUserPreferencesWithHttpInfo($user_id)
     {
-        
         // verify the required parameter 'user_id' is set
         if ($user_id === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_id when calling deleteUserPreferences');
         }
-  
         // parse inputs
         $resourcePath = "/{userId}";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
-        
-        
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // path params
-        
         if ($user_id !== null) {
             $resourcePath = str_replace(
                 "{" . "userId" . "}",
@@ -151,77 +152,71 @@ class UserPreferencesApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-AccessToken');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-AccessToken'] = $apiKey;
         }
-        
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-UserId');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-UserId'] = $apiKey;
         }
-        
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'DELETE',
-                $queryParams, $httpBody,
-                $headerParams
+                $resourcePath,
+                'DELETE',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                null,
+                '/{userId}'
             );
-            
+
             return array(null, $statusCode, $httpHeader);
-            
         } catch (ApiException $e) {
-            switch ($e->getCode()) { 
+            switch ($e->getCode()) {
             }
-  
+
             throw $e;
         }
     }
-    
+
     /**
-     * getGlobalUserPreferenceByName
+     * Operation getGlobalUserPreferenceByName
      *
      * Returns single specific global preference for specific user
      *
      * @param string $user_id The Id of the user whose preference will be returned (required)
-     * @param string $preference_name The name of preference whose value will be returned  (required)
+     * @param string $preference_name The name of preference whose value will be returned (required)
      * @return \Swagger\Client\User\Preferences\Models\GlobalPreference
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function getGlobalUserPreferenceByName($user_id, $preference_name)
     {
-        list($response, $statusCode, $httpHeader) = $this->getGlobalUserPreferenceByNameWithHttpInfo ($user_id, $preference_name);
-        return $response; 
+        list($response) = $this->getGlobalUserPreferenceByNameWithHttpInfo($user_id, $preference_name);
+        return $response;
     }
 
-
     /**
-     * getGlobalUserPreferenceByNameWithHttpInfo
+     * Operation getGlobalUserPreferenceByNameWithHttpInfo
      *
      * Returns single specific global preference for specific user
      *
      * @param string $user_id The Id of the user whose preference will be returned (required)
-     * @param string $preference_name The name of preference whose value will be returned  (required)
+     * @param string $preference_name The name of preference whose value will be returned (required)
      * @return Array of \Swagger\Client\User\Preferences\Models\GlobalPreference, HTTP status code, HTTP response headers (array of strings)
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function getGlobalUserPreferenceByNameWithHttpInfo($user_id, $preference_name)
     {
-        
         // verify the required parameter 'user_id' is set
         if ($user_id === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_id when calling getGlobalUserPreferenceByName');
@@ -230,31 +225,27 @@ class UserPreferencesApi
         if ($preference_name === null) {
             throw new \InvalidArgumentException('Missing the required parameter $preference_name when calling getGlobalUserPreferenceByName');
         }
-  
         // parse inputs
         $resourcePath = "/{userId}/global/{preferenceName}";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
-        
-        
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // path params
-        
         if ($user_id !== null) {
             $resourcePath = str_replace(
                 "{" . "userId" . "}",
                 $this->apiClient->getSerializer()->toPathValue($user_id),
                 $resourcePath
             );
-        }// path params
-        
+        }
+        // path params
         if ($preference_name !== null) {
             $resourcePath = str_replace(
                 "{" . "preferenceName" . "}",
@@ -266,91 +257,81 @@ class UserPreferencesApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-AccessToken');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-AccessToken'] = $apiKey;
         }
-        
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-UserId');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-UserId'] = $apiKey;
         }
-        
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'GET',
-                $queryParams, $httpBody,
-                $headerParams, '\Swagger\Client\User\Preferences\Models\GlobalPreference'
+                $resourcePath,
+                'GET',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                '\Swagger\Client\User\Preferences\Models\GlobalPreference',
+                '/{userId}/global/{preferenceName}'
             );
-            
-            if (!$response) {
-                return array(null, $statusCode, $httpHeader);
+
+            return array($this->apiClient->getSerializer()->deserialize($response, '\Swagger\Client\User\Preferences\Models\GlobalPreference', $httpHeader), $statusCode, $httpHeader);
+        } catch (ApiException $e) {
+            switch ($e->getCode()) {
+                case 200:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\GlobalPreference', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
+                case 404:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
 
-            return array(\Swagger\Client\ObjectSerializer::deserialize($response, '\Swagger\Client\User\Preferences\Models\GlobalPreference', $httpHeader), $statusCode, $httpHeader);
-            
-        } catch (ApiException $e) {
-            switch ($e->getCode()) { 
-            case 200:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\GlobalPreference', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            case 404:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            }
-  
             throw $e;
         }
     }
-    
+
     /**
-     * getLocalUserPreferenceByName
+     * Operation getLocalUserPreferenceByName
      *
      * Returns single specific local preference for specific user
      *
      * @param string $user_id The Id of the user whose preference will be returned (required)
      * @param int $wiki_id The Id of the wiki for which preference will be returned (required)
-     * @param string $preference_name The name of preference whose value will be returned  (required)
+     * @param string $preference_name The name of preference whose value will be returned (required)
      * @return \Swagger\Client\User\Preferences\Models\LocalPreference
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function getLocalUserPreferenceByName($user_id, $wiki_id, $preference_name)
     {
-        list($response, $statusCode, $httpHeader) = $this->getLocalUserPreferenceByNameWithHttpInfo ($user_id, $wiki_id, $preference_name);
-        return $response; 
+        list($response) = $this->getLocalUserPreferenceByNameWithHttpInfo($user_id, $wiki_id, $preference_name);
+        return $response;
     }
 
-
     /**
-     * getLocalUserPreferenceByNameWithHttpInfo
+     * Operation getLocalUserPreferenceByNameWithHttpInfo
      *
      * Returns single specific local preference for specific user
      *
      * @param string $user_id The Id of the user whose preference will be returned (required)
      * @param int $wiki_id The Id of the wiki for which preference will be returned (required)
-     * @param string $preference_name The name of preference whose value will be returned  (required)
+     * @param string $preference_name The name of preference whose value will be returned (required)
      * @return Array of \Swagger\Client\User\Preferences\Models\LocalPreference, HTTP status code, HTTP response headers (array of strings)
      * @throws \Swagger\Client\ApiException on non-2xx response
      */
     public function getLocalUserPreferenceByNameWithHttpInfo($user_id, $wiki_id, $preference_name)
     {
-        
         // verify the required parameter 'user_id' is set
         if ($user_id === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_id when calling getLocalUserPreferenceByName');
@@ -363,39 +344,35 @@ class UserPreferencesApi
         if ($preference_name === null) {
             throw new \InvalidArgumentException('Missing the required parameter $preference_name when calling getLocalUserPreferenceByName');
         }
-  
         // parse inputs
         $resourcePath = "/{userId}/local/{preferenceName}/wikia/{wikiId}";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
-        
-        
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // path params
-        
         if ($user_id !== null) {
             $resourcePath = str_replace(
                 "{" . "userId" . "}",
                 $this->apiClient->getSerializer()->toPathValue($user_id),
                 $resourcePath
             );
-        }// path params
-        
+        }
+        // path params
         if ($wiki_id !== null) {
             $resourcePath = str_replace(
                 "{" . "wikiId" . "}",
                 $this->apiClient->getSerializer()->toPathValue($wiki_id),
                 $resourcePath
             );
-        }// path params
-        
+        }
+        // path params
         if ($preference_name !== null) {
             $resourcePath = str_replace(
                 "{" . "preferenceName" . "}",
@@ -407,61 +384,53 @@ class UserPreferencesApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-AccessToken');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-AccessToken'] = $apiKey;
         }
-        
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-UserId');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-UserId'] = $apiKey;
         }
-        
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'GET',
-                $queryParams, $httpBody,
-                $headerParams, '\Swagger\Client\User\Preferences\Models\LocalPreference'
+                $resourcePath,
+                'GET',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                '\Swagger\Client\User\Preferences\Models\LocalPreference',
+                '/{userId}/local/{preferenceName}/wikia/{wikiId}'
             );
-            
-            if (!$response) {
-                return array(null, $statusCode, $httpHeader);
+
+            return array($this->apiClient->getSerializer()->deserialize($response, '\Swagger\Client\User\Preferences\Models\LocalPreference', $httpHeader), $statusCode, $httpHeader);
+        } catch (ApiException $e) {
+            switch ($e->getCode()) {
+                case 200:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\LocalPreference', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
+                case 404:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
 
-            return array(\Swagger\Client\ObjectSerializer::deserialize($response, '\Swagger\Client\User\Preferences\Models\LocalPreference', $httpHeader), $statusCode, $httpHeader);
-            
-        } catch (ApiException $e) {
-            switch ($e->getCode()) { 
-            case 200:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\LocalPreference', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            case 404:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            }
-  
             throw $e;
         }
     }
-    
+
     /**
-     * getUserPreferences
+     * Operation getUserPreferences
      *
      * Returns all the global user preferences for a user
      *
@@ -471,13 +440,12 @@ class UserPreferencesApi
      */
     public function getUserPreferences($user_id)
     {
-        list($response, $statusCode, $httpHeader) = $this->getUserPreferencesWithHttpInfo ($user_id);
-        return $response; 
+        list($response) = $this->getUserPreferencesWithHttpInfo($user_id);
+        return $response;
     }
 
-
     /**
-     * getUserPreferencesWithHttpInfo
+     * Operation getUserPreferencesWithHttpInfo
      *
      * Returns all the global user preferences for a user
      *
@@ -487,28 +455,23 @@ class UserPreferencesApi
      */
     public function getUserPreferencesWithHttpInfo($user_id)
     {
-        
         // verify the required parameter 'user_id' is set
         if ($user_id === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_id when calling getUserPreferences');
         }
-  
         // parse inputs
         $resourcePath = "/{userId}";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
-        
-        
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // path params
-        
         if ($user_id !== null) {
             $resourcePath = str_replace(
                 "{" . "userId" . "}",
@@ -520,61 +483,53 @@ class UserPreferencesApi
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
         
-        
-  
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-AccessToken');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-AccessToken'] = $apiKey;
         }
-        
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-UserId');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-UserId'] = $apiKey;
         }
-        
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'GET',
-                $queryParams, $httpBody,
-                $headerParams, '\Swagger\Client\User\Preferences\Models\UserPreferences'
+                $resourcePath,
+                'GET',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                '\Swagger\Client\User\Preferences\Models\UserPreferences',
+                '/{userId}'
             );
-            
-            if (!$response) {
-                return array(null, $statusCode, $httpHeader);
+
+            return array($this->apiClient->getSerializer()->deserialize($response, '\Swagger\Client\User\Preferences\Models\UserPreferences', $httpHeader), $statusCode, $httpHeader);
+        } catch (ApiException $e) {
+            switch ($e->getCode()) {
+                case 200:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\UserPreferences', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
+                case 404:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
 
-            return array(\Swagger\Client\ObjectSerializer::deserialize($response, '\Swagger\Client\User\Preferences\Models\UserPreferences', $httpHeader), $statusCode, $httpHeader);
-            
-        } catch (ApiException $e) {
-            switch ($e->getCode()) { 
-            case 200:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\UserPreferences', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            case 404:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
-            }
-  
             throw $e;
         }
     }
-    
+
     /**
-     * setUserPreferences
+     * Operation setUserPreferences
      *
      * set a user's preferences. note - this deletes preferences not in the provided list.
      *
@@ -585,13 +540,12 @@ class UserPreferencesApi
      */
     public function setUserPreferences($user_id, $user_preferences)
     {
-        list($response, $statusCode, $httpHeader) = $this->setUserPreferencesWithHttpInfo ($user_id, $user_preferences);
-        return $response; 
+        list($response) = $this->setUserPreferencesWithHttpInfo($user_id, $user_preferences);
+        return $response;
     }
 
-
     /**
-     * setUserPreferencesWithHttpInfo
+     * Operation setUserPreferencesWithHttpInfo
      *
      * set a user's preferences. note - this deletes preferences not in the provided list.
      *
@@ -602,7 +556,6 @@ class UserPreferencesApi
      */
     public function setUserPreferencesWithHttpInfo($user_id, $user_preferences)
     {
-        
         // verify the required parameter 'user_id' is set
         if ($user_id === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_id when calling setUserPreferences');
@@ -611,23 +564,19 @@ class UserPreferencesApi
         if ($user_preferences === null) {
             throw new \InvalidArgumentException('Missing the required parameter $user_preferences when calling setUserPreferences');
         }
-  
         // parse inputs
         $resourcePath = "/{userId}";
         $httpBody = '';
         $queryParams = array();
         $headerParams = array();
         $formParams = array();
-        $_header_accept = ApiClient::selectHeaderAccept(array('application/json'));
+        $_header_accept = $this->apiClient->selectHeaderAccept(array('application/json'));
         if (!is_null($_header_accept)) {
             $headerParams['Accept'] = $_header_accept;
         }
-        $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array());
-  
-        
-        
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(array());
+
         // path params
-        
         if ($user_id !== null) {
             $resourcePath = str_replace(
                 "{" . "userId" . "}",
@@ -638,54 +587,51 @@ class UserPreferencesApi
         // default format to json
         $resourcePath = str_replace("{format}", "json", $resourcePath);
 
-        
         // body params
         $_tempBody = null;
         if (isset($user_preferences)) {
             $_tempBody = $user_preferences;
         }
-  
+
         // for model (json/xml)
         if (isset($_tempBody)) {
             $httpBody = $_tempBody; // $_tempBody is the method argument, if present
         } elseif (count($formParams) > 0) {
             $httpBody = $formParams; // for HTTP post (form)
         }
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-AccessToken');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-AccessToken'] = $apiKey;
         }
-        
-        
         // this endpoint requires API key authentication
         $apiKey = $this->apiClient->getApiKeyWithPrefix('X-Wikia-UserId');
         if (strlen($apiKey) !== 0) {
             $headerParams['X-Wikia-UserId'] = $apiKey;
         }
-        
-        
         // make the API Call
         try {
             list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
-                $resourcePath, 'PUT',
-                $queryParams, $httpBody,
-                $headerParams
+                $resourcePath,
+                'PUT',
+                $queryParams,
+                $httpBody,
+                $headerParams,
+                null,
+                '/{userId}'
             );
-            
+
             return array(null, $statusCode, $httpHeader);
-            
         } catch (ApiException $e) {
-            switch ($e->getCode()) { 
-            case 400:
-                $data = \Swagger\Client\ObjectSerializer::deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
-                $e->setResponseObject($data);
-                break;
+            switch ($e->getCode()) {
+                case 400:
+                    $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Swagger\Client\User\Preferences\Models\Problem', $e->getResponseHeaders());
+                    $e->setResponseObject($data);
+                    break;
             }
-  
+
             throw $e;
         }
     }
-    
+
 }
