@@ -10,21 +10,25 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -34,11 +38,12 @@
 namespace Swagger\Client\User\Preferences\Models;
 
 use \ArrayAccess;
+
 /**
  * Problem Class Doc Comment
  *
- * @category    Class
- * @description 
+ * @category    Class */
+/** 
  * @package     Swagger\Client
  * @author      http://github.com/swagger-api/swagger-codegen
  * @license     http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
@@ -47,100 +52,88 @@ use \ArrayAccess;
 class Problem implements ArrayAccess
 {
     /**
-      * Array of property to type mappings. Used for (de)serialization 
+      * The original name of the model.
+      * @var string
+      */
+    protected static $swaggerModelName = 'Problem';
+
+    /**
+      * Array of property to type mappings. Used for (de)serialization
       * @var string[]
       */
-    static $swaggerTypes = array(
+    protected static $swaggerTypes = array(
         'title' => 'string',
         'status' => 'int',
         'type' => 'string',
         'detail' => 'string',
         'instance' => 'string'
     );
-  
-    static function swaggerTypes() {
+
+    public static function swaggerTypes()
+    {
         return self::$swaggerTypes;
     }
 
-    /** 
-      * Array of attributes where the key is the local name, and the value is the original name
-      * @var string[] 
-      */
-    static $attributeMap = array(
+    /**
+     * Array of attributes where the key is the local name, and the value is the original name
+     * @var string[]
+     */
+    protected static $attributeMap = array(
         'title' => 'title',
         'status' => 'status',
         'type' => 'type',
         'detail' => 'detail',
         'instance' => 'instance'
     );
-  
-    static function attributeMap() {
+
+    public static function attributeMap()
+    {
         return self::$attributeMap;
     }
 
     /**
-      * Array of attributes to setter functions (for deserialization of responses)
-      * @var string[]
-      */
-    static $setters = array(
+     * Array of attributes to setter functions (for deserialization of responses)
+     * @var string[]
+     */
+    protected static $setters = array(
         'title' => 'setTitle',
         'status' => 'setStatus',
         'type' => 'setType',
         'detail' => 'setDetail',
         'instance' => 'setInstance'
     );
-  
-    static function setters() {
+
+    public static function setters()
+    {
         return self::$setters;
     }
 
     /**
-      * Array of attributes to getter functions (for serialization of requests)
-      * @var string[]
-      */
-    static $getters = array(
+     * Array of attributes to getter functions (for serialization of requests)
+     * @var string[]
+     */
+    protected static $getters = array(
         'title' => 'getTitle',
         'status' => 'getStatus',
         'type' => 'getType',
         'detail' => 'getDetail',
         'instance' => 'getInstance'
     );
-  
-    static function getters() {
+
+    public static function getters()
+    {
         return self::$getters;
     }
 
     
-    /**
-      * $title 
-      * @var string
-      */
-    protected $title;
+
     
+
     /**
-      * $status 
-      * @var int
-      */
-    protected $status;
-    
-    /**
-      * $type 
-      * @var string
-      */
-    protected $type;
-    
-    /**
-      * $detail 
-      * @var string
-      */
-    protected $detail;
-    
-    /**
-      * $instance 
-      * @var string
-      */
-    protected $instance;
-    
+     * Associative array for storing property values
+     * @var mixed[]
+     */
+    protected $container = array();
 
     /**
      * Constructor
@@ -148,172 +141,209 @@ class Problem implements ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        
-        if ($data != null) {
-            $this->title = $data["title"];
-            $this->status = $data["status"];
-            $this->type = $data["type"];
-            $this->detail = $data["detail"];
-            $this->instance = $data["instance"];
-        }
+        $this->container['title'] = isset($data['title']) ? $data['title'] : null;
+        $this->container['status'] = isset($data['status']) ? $data['status'] : null;
+        $this->container['type'] = isset($data['type']) ? $data['type'] : null;
+        $this->container['detail'] = isset($data['detail']) ? $data['detail'] : null;
+        $this->container['instance'] = isset($data['instance']) ? $data['instance'] : null;
     }
-    
+
+    /**
+     * show all the invalid properties with reasons.
+     *
+     * @return array invalid properties with reasons
+     */
+    public function listInvalidProperties()
+    {
+        $invalid_properties = array();
+        if ($this->container['title'] === null) {
+            $invalid_properties[] = "'title' can't be null";
+        }
+        if ($this->container['status'] === null) {
+            $invalid_properties[] = "'status' can't be null";
+        }
+        return $invalid_properties;
+    }
+
+    /**
+     * validate all the properties in the model
+     * return true if all passed
+     *
+     * @return bool True if all properteis are valid
+     */
+    public function valid()
+    {
+        if ($this->container['title'] === null) {
+            return false;
+        }
+        if ($this->container['status'] === null) {
+            return false;
+        }
+        return true;
+    }
+
+
     /**
      * Gets title
      * @return string
      */
     public function getTitle()
     {
-        return $this->title;
+        return $this->container['title'];
     }
-  
+
     /**
      * Sets title
-     * @param string $title 
+     * @param string $title
      * @return $this
      */
     public function setTitle($title)
     {
-        
-        $this->title = $title;
+        $this->container['title'] = $title;
+
         return $this;
     }
-    
+
     /**
      * Gets status
      * @return int
      */
     public function getStatus()
     {
-        return $this->status;
+        return $this->container['status'];
     }
-  
+
     /**
      * Sets status
-     * @param int $status 
+     * @param int $status
      * @return $this
      */
     public function setStatus($status)
     {
-        
-        $this->status = $status;
+        $this->container['status'] = $status;
+
         return $this;
     }
-    
+
     /**
      * Gets type
      * @return string
      */
     public function getType()
     {
-        return $this->type;
+        return $this->container['type'];
     }
-  
+
     /**
      * Sets type
-     * @param string $type 
+     * @param string $type
      * @return $this
      */
     public function setType($type)
     {
-        
-        $this->type = $type;
+        $this->container['type'] = $type;
+
         return $this;
     }
-    
+
     /**
      * Gets detail
      * @return string
      */
     public function getDetail()
     {
-        return $this->detail;
+        return $this->container['detail'];
     }
-  
+
     /**
      * Sets detail
-     * @param string $detail 
+     * @param string $detail
      * @return $this
      */
     public function setDetail($detail)
     {
-        
-        $this->detail = $detail;
+        $this->container['detail'] = $detail;
+
         return $this;
     }
-    
+
     /**
      * Gets instance
      * @return string
      */
     public function getInstance()
     {
-        return $this->instance;
+        return $this->container['instance'];
     }
-  
+
     /**
      * Sets instance
-     * @param string $instance 
+     * @param string $instance
      * @return $this
      */
     public function setInstance($instance)
     {
-        
-        $this->instance = $instance;
+        $this->container['instance'] = $instance;
+
         return $this;
     }
-    
     /**
      * Returns true if offset exists. False otherwise.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @return boolean
      */
     public function offsetExists($offset)
     {
-        return isset($this->$offset);
+        return isset($this->container[$offset]);
     }
-  
+
     /**
      * Gets offset.
-     * @param  integer $offset Offset 
-     * @return mixed 
+     * @param  integer $offset Offset
+     * @return mixed
      */
     public function offsetGet($offset)
     {
-        return $this->$offset;
+        return isset($this->container[$offset]) ? $this->container[$offset] : null;
     }
-  
+
     /**
      * Sets value based on offset.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @param  mixed   $value  Value to be set
      * @return void
      */
     public function offsetSet($offset, $value)
     {
-        $this->$offset = $value;
+        if (is_null($offset)) {
+            $this->container[] = $value;
+        } else {
+            $this->container[$offset] = $value;
+        }
     }
-  
+
     /**
      * Unsets offset.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @return void
      */
     public function offsetUnset($offset)
     {
-        unset($this->$offset);
+        unset($this->container[$offset]);
     }
-  
+
     /**
      * Gets the string presentation of the object
      * @return string
      */
     public function __toString()
     {
-        if (defined('JSON_PRETTY_PRINT')) {
+        if (defined('JSON_PRETTY_PRINT')) { // use JSON pretty print
             return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this), JSON_PRETTY_PRINT);
-        } else {
-            return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this));
         }
+
+        return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this));
     }
 }
+
+
