@@ -10,21 +10,25 @@
  * @license  http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
  * @link     https://github.com/swagger-api/swagger-codegen
  */
+
 /**
- *  Copyright 2016 SmartBear Software
+ * 
+ * 
+ * Generated by: https://github.com/swagger-api/swagger-codegen.git
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -34,11 +38,12 @@
 namespace Swagger\Client\User\Preferences\Models;
 
 use \ArrayAccess;
+
 /**
  * UserPreferences Class Doc Comment
  *
- * @category    Class
- * @description 
+ * @category    Class */
+/** 
  * @package     Swagger\Client
  * @author      http://github.com/swagger-api/swagger-codegen
  * @license     http://www.apache.org/licenses/LICENSE-2.0 Apache Licene v2
@@ -47,70 +52,76 @@ use \ArrayAccess;
 class UserPreferences implements ArrayAccess
 {
     /**
-      * Array of property to type mappings. Used for (de)serialization 
+      * The original name of the model.
+      * @var string
+      */
+    protected static $swaggerModelName = 'UserPreferences';
+
+    /**
+      * Array of property to type mappings. Used for (de)serialization
       * @var string[]
       */
-    static $swaggerTypes = array(
+    protected static $swaggerTypes = array(
         'global_preferences' => '\Swagger\Client\User\Preferences\Models\GlobalPreference[]',
         'local_preferences' => '\Swagger\Client\User\Preferences\Models\LocalPreference[]'
     );
-  
-    static function swaggerTypes() {
+
+    public static function swaggerTypes()
+    {
         return self::$swaggerTypes;
     }
 
-    /** 
-      * Array of attributes where the key is the local name, and the value is the original name
-      * @var string[] 
-      */
-    static $attributeMap = array(
+    /**
+     * Array of attributes where the key is the local name, and the value is the original name
+     * @var string[]
+     */
+    protected static $attributeMap = array(
         'global_preferences' => 'globalPreferences',
         'local_preferences' => 'localPreferences'
     );
-  
-    static function attributeMap() {
+
+    public static function attributeMap()
+    {
         return self::$attributeMap;
     }
 
     /**
-      * Array of attributes to setter functions (for deserialization of responses)
-      * @var string[]
-      */
-    static $setters = array(
+     * Array of attributes to setter functions (for deserialization of responses)
+     * @var string[]
+     */
+    protected static $setters = array(
         'global_preferences' => 'setGlobalPreferences',
         'local_preferences' => 'setLocalPreferences'
     );
-  
-    static function setters() {
+
+    public static function setters()
+    {
         return self::$setters;
     }
 
     /**
-      * Array of attributes to getter functions (for serialization of requests)
-      * @var string[]
-      */
-    static $getters = array(
+     * Array of attributes to getter functions (for serialization of requests)
+     * @var string[]
+     */
+    protected static $getters = array(
         'global_preferences' => 'getGlobalPreferences',
         'local_preferences' => 'getLocalPreferences'
     );
-  
-    static function getters() {
+
+    public static function getters()
+    {
         return self::$getters;
     }
 
     
-    /**
-      * $global_preferences 
-      * @var \Swagger\Client\User\Preferences\Models\GlobalPreference[]
-      */
-    protected $global_preferences;
+
     
+
     /**
-      * $local_preferences 
-      * @var \Swagger\Client\User\Preferences\Models\LocalPreference[]
-      */
-    protected $local_preferences;
-    
+     * Associative array for storing property values
+     * @var mixed[]
+     */
+    protected $container = array();
 
     /**
      * Constructor
@@ -118,106 +129,131 @@ class UserPreferences implements ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        
-        if ($data != null) {
-            $this->global_preferences = $data["global_preferences"];
-            $this->local_preferences = $data["local_preferences"];
-        }
+        $this->container['global_preferences'] = isset($data['global_preferences']) ? $data['global_preferences'] : null;
+        $this->container['local_preferences'] = isset($data['local_preferences']) ? $data['local_preferences'] : null;
     }
-    
+
+    /**
+     * show all the invalid properties with reasons.
+     *
+     * @return array invalid properties with reasons
+     */
+    public function listInvalidProperties()
+    {
+        $invalid_properties = array();
+        return $invalid_properties;
+    }
+
+    /**
+     * validate all the properties in the model
+     * return true if all passed
+     *
+     * @return bool True if all properteis are valid
+     */
+    public function valid()
+    {
+        return true;
+    }
+
+
     /**
      * Gets global_preferences
      * @return \Swagger\Client\User\Preferences\Models\GlobalPreference[]
      */
     public function getGlobalPreferences()
     {
-        return $this->global_preferences;
+        return $this->container['global_preferences'];
     }
-  
+
     /**
      * Sets global_preferences
-     * @param \Swagger\Client\User\Preferences\Models\GlobalPreference[] $global_preferences 
+     * @param \Swagger\Client\User\Preferences\Models\GlobalPreference[] $global_preferences
      * @return $this
      */
     public function setGlobalPreferences($global_preferences)
     {
-        
-        $this->global_preferences = $global_preferences;
+        $this->container['global_preferences'] = $global_preferences;
+
         return $this;
     }
-    
+
     /**
      * Gets local_preferences
      * @return \Swagger\Client\User\Preferences\Models\LocalPreference[]
      */
     public function getLocalPreferences()
     {
-        return $this->local_preferences;
+        return $this->container['local_preferences'];
     }
-  
+
     /**
      * Sets local_preferences
-     * @param \Swagger\Client\User\Preferences\Models\LocalPreference[] $local_preferences 
+     * @param \Swagger\Client\User\Preferences\Models\LocalPreference[] $local_preferences
      * @return $this
      */
     public function setLocalPreferences($local_preferences)
     {
-        
-        $this->local_preferences = $local_preferences;
+        $this->container['local_preferences'] = $local_preferences;
+
         return $this;
     }
-    
     /**
      * Returns true if offset exists. False otherwise.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @return boolean
      */
     public function offsetExists($offset)
     {
-        return isset($this->$offset);
+        return isset($this->container[$offset]);
     }
-  
+
     /**
      * Gets offset.
-     * @param  integer $offset Offset 
-     * @return mixed 
+     * @param  integer $offset Offset
+     * @return mixed
      */
     public function offsetGet($offset)
     {
-        return $this->$offset;
+        return isset($this->container[$offset]) ? $this->container[$offset] : null;
     }
-  
+
     /**
      * Sets value based on offset.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @param  mixed   $value  Value to be set
      * @return void
      */
     public function offsetSet($offset, $value)
     {
-        $this->$offset = $value;
+        if (is_null($offset)) {
+            $this->container[] = $value;
+        } else {
+            $this->container[$offset] = $value;
+        }
     }
-  
+
     /**
      * Unsets offset.
-     * @param  integer $offset Offset 
+     * @param  integer $offset Offset
      * @return void
      */
     public function offsetUnset($offset)
     {
-        unset($this->$offset);
+        unset($this->container[$offset]);
     }
-  
+
     /**
      * Gets the string presentation of the object
      * @return string
      */
     public function __toString()
     {
-        if (defined('JSON_PRETTY_PRINT')) {
+        if (defined('JSON_PRETTY_PRINT')) { // use JSON pretty print
             return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this), JSON_PRETTY_PRINT);
-        } else {
-            return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this));
         }
+
+        return json_encode(\Swagger\Client\ObjectSerializer::sanitizeForSerialization($this));
     }
 }
+
+
